### 2024-11-12 #7448 @mncrowe
> @mncrowe I am going to ping our editorial board to see if this submission passes our criteria for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This process will take a few weeks; thanks for your patience.

Thank you for your consideration. It appears that the query-scope label almost always leads to rejection, but if it does help my case, I'd like to point out that much of the commit history was lost due to a force push on 04/11/2024. The commit count above only seems to include the commits after this point whereas the full [activity](https://github.com/mncrowe/QGDipoles.jl/activity?sort=ASC) goes back to 28/06/2024.
