### 2024-11-12 31ded9d
Update paper.bib with missing DOI fields
