+++ paper/paper.bib
@@ -56,7 +56,7 @@
           	 Andr{{\'e}} Pal{{\'o}}czy},
   title = 	{GeophysicalFlows.jl: Solvers for geophysical fluid dynamics problems in periodic domains on CPUs \& GPUs},
   journal = 	{Journal of Open Source Software},
-  doi =		{https://doi.org/10.21105/joss.03053}
+  doi =		{10.21105/joss.03053}
 }
 
 @article{Crowe_Johnson_23,
@@ -66,7 +66,8 @@
   journal   = {Journal of Fluid Mechanics},
   year      = {2023},
   volume    = {970},
-  pages     = {A10}
+  pages     = {A10},
+  doi       = {10.1017/jfm.2023.607}
 }
 
 @article{Crowe_Johnson_2024,
@@ -76,7 +77,8 @@
   author =   {Crowe, Matthew N. and
               Johnson, Edward R.},
   year =     {2024},
-  pages =    {R1}
+  pages =    {R1},
+  doi =      {10.1017/jfm.2024.619}
 }
 
 @Article{Johnson_Crowe_23,
@@ -87,6 +89,7 @@
   year           = {2023},
   volume         = {958},
   pages          = {R2},
+  doi            = {10.1017/jfm.2023.87},
 }
 
 @Article{NiZWH20,
@@ -99,6 +102,7 @@
   volume    = {125},
   journal   = {Journal of Geophysical Research: Oceans},
   year      = {2020},
+  doi       = {10.1029/2020jc016479}
 }
 
 @Article{NycanderI90,
@@ -109,6 +113,7 @@
   volume  = {2},
   journal = {Physics of Fluids},
   year    = {1990},
+  doi     = {10.1063/1.859425}
 }
 
 @article{Davies_et_al,
@@ -122,6 +127,7 @@
   number  = {11},
   pages   = {116601},
   year    = {2023},
+  doi     = {10.1063/5.0171909}
 }
 
 @article {mcwilliams_1980,
@@ -130,7 +136,8 @@
   journal  = {Dynamics of Atmospheres and Oceans},
   year     = {1980},
   volume   = {5},
-  pages    = {43-66}
+  pages    = {43-66},
+  doi      = {10.1016/0377-0265(80)90010-x}
 }
 
 @article{rostami_zeitlin_2021,
@@ -142,6 +149,7 @@
   number   = {3},
   pages    = {345-367},
   year     = {2021},
+  doi      = {10.1080/03091929.2020.1805448}
 }
 
 @article{Dedalus,
