### 2024-11-12 #7448 @mncrowe
>```
> ❌ MISSING DOIs
>
> - 10.1017/jfm.2023.607 may be a valid DOI for title: The evolution of surface quasi-geostrophic modons ...
> - 10.1017/jfm.2024.619 may be a valid DOI for title: Modon solutions in an N-layer quasi-geostrophic mo...
> - 10.1017/jfm.2023.87 may be a valid DOI for title: Oceanic dipoles in a surface quasi-geostrophic mod...
> - 10.1029/2020jc016479 may be a valid DOI for title: Widespread Mesoscale Dipoles in the Global Ocean
> - 10.1063/1.859425 may be a valid DOI for title: Motion of dipole vortices in a weakly inhomogeneou...
> - 10.1063/5.0171909 may be a valid DOI for title: Deformation and destruction of north-eastward drif...
> - 10.1016/0377-0265(80)90010-x may be a valid DOI for title: An application of equivalent modons to atmospheric...
> - 10.1080/03091929.2020.1805448 may be a valid DOI for title: Eastward-moving equatorial modons in moist-convect...
>
> ❌ INVALID DOIs
>
> - https://doi.org/10.21105/joss.03053 is INVALID because of 'https://doi.org/' prefix
> ```

I've fixed the missing and invalid DOIs.
