### 2024-11-12 502bf92
added plots to example files, started work on monopoles script
