### 2024-11-12 f8b300f
added axis labels to docs/src/Examples.md
