+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"Examples/#Examples","page":"Examples","title":"Examples","text":"","category":"section"},{"location":"Examples/","page":"Examples","title":"Examples","text":"Here we present some examples which demonstrate how to use this package. Further examples are available in the examples/ directory here.","category":"page"},{"location":"Examples/#Example-1:-1-layer-QG","page":"Examples","title":"Example 1: 1-layer QG","text":"","category":"section"},{"location":"Examples/","page":"Examples","title":"Examples","text":"Let's calculate and plot the Larichev-Reznik dipole (LRD). This diople exists on the beta-plane in the equivalent barotropic model so we take beta = R = 1 and consider a 1-layer solution (N = 1). We'll also assume unit radius and velocity, ell = U = 1. Let's start by loading the package and defining some parameters.","category":"page"},{"location":"Examples/","page":"Examples","title":"Examples","text":"using QGDipoles\n\n# Set problem parameters\n\nU, ℓ = 1, 1\t# vortex speed and radius\nR = 1\t\t# Rossby radius in each layer\nβ = 1\t\t# background PV gradient in each layer\n\nM = 8\t\t# number of coefficients in Zernike expansion\ntol = 1e-8\t# maximum error in solution evaluation\n\n# Set grid parameters\n\nNx, Ny = 512, 512\nLx, Ly = 10, 10\n\nnothing","category":"page"},{"location":"Examples/","page":"Examples","title":"Examples","text":"We've taken M = 8 as this is generally a sufficient number of terms to get a relative error  10^-6 in the final result. The tolerance, tol, is used in calculating the terms in the linear system and a value of tol=10^{-8} corresponds to approximately the same error as our chosen M value. We're also going to build a grid with 512 points in each direction and have taken the grid size to be 10 in each direction, which is sufficient to capture the far-field decay of the vortex. We can now build the linear system and solve for the coefficients as follows:","category":"page"},{"location":"Examples/","page":"Examples","title":"Examples","text":"# Build and solve linear system for coefficients\n\nλ = ℓ / R\nμ = β * ℓ^2/U\n\nA, B, c, d = BuildLinSys(M, λ, μ; tol)\nK, a = SolveInhomEVP(A, B, c, d; K₀ = 4, tol)\n\nnothing","category":"page"},{"location":"Examples/","page":"Examples","title":"Examples","text":"The intermediate parameters, lambda and mu, describe the rescaled vortex radius and PV gradient. Finally, we can define a grid and evaluate our streamfunction, PV and velocities using:","category":"page"},{"location":"Examples/","page":"Examples","title":"Examples","text":"# Create grid and calculate streamfunctions and vorticities\n\ngrid = CreateGrid(Nx, Ny, Lx, Ly)\nψ, q = Calc_ψq(a, U, ℓ, R, β, grid)\nu, v = Calc_uv(ψ, grid)\n\nnothing","category":"page"},{"location":"Examples/","page":"Examples","title":"Examples","text":"We can plot our solution using Plots.jl:","category":"page"},{"location":"Examples/","page":"Examples","title":"Examples","text":"using Plots\n\nheatmap(grid.x, grid.y, transpose(ψ[:,:,1]);\n    colormap = :balance,\n    aspect_ratio=1,\n    xlims = (-Lx/2, Lx/2),\n    ylims = (-Ly/2, Ly/2))\n","category":"page"},{"location":"Examples/","page":"Examples","title":"Examples","text":"Note that we transpose psi when plotting as x corresonds to the first dimension of psi.","category":"page"},{"location":"Examples/#Example-2:-multi-layer-QG","page":"Examples","title":"Example
[…]
