### 2024-11-12 8e4256d
fixed bug with 3-layer example figure axes, added multi-panel figures, replaced figures for example 5 with single svg
