+++ examples/2_Layer_Modon.jl
@@ -30,14 +30,21 @@ K, a = SolveInhomEVP(A, B, c, d; tol)
 grid = CreateGrid(Nx, Ny, Lx, Ly; cuda)
 ψ, q = Calc_ψq(a, U, ℓ, R, β, grid)
 
-# Plot streamfunction ψ in layer 1
+# Plot streamfunction ψ in both layers
 
 using Plots
 
-heatmap(grid.x, grid.y, transpose(ψ[:, :, 1]);
+plot(heatmap(grid.x, grid.y, transpose(ψ[:, :, 1]);
 	colormap = :balance,
 	aspect_ratio = 1,
 	xlims = (-Lx/2, Lx/2),
 	ylims = (-Ly/2, Ly/2),
 	xlabel = "x",
-	ylabel = "y")
+	ylabel = "y"),
+     heatmap(grid.x, grid.y, transpose(ψ[:, :, 2]);
+	colormap = :balance,
+	aspect_ratio = 1,
+	xlims = (-Lx/2, Lx/2),
+	ylims = (-Ly/2, Ly/2),
+	xlabel = "x",
+	ylabel = "y"))
