### 2024-11-12 bfeab65
fixed typo, png -> svg
