### 2024-11-12 a990e9f
expand tests to cover GPU functionality
