### 2024-11-12 c317741
added tests for wrapper functions and for active/passive layer functionality
