### 2024-11-12 339838b
Tidy up module file and unwrap nested functions in JJ_int.jl
