### 2024-11-12 ce71983
Tidy up lin_sys.jl and fix test bug with TestLRD (tolerance too low)
