+++ src/lin_sys.jl
@@ -74,9 +74,10 @@ function BuildLinSys(M::Int, λ::Union{Vector,Number}, μ::Union{Vector,Number};
 
 		# Define A, B, c, d for the SQG system, B is temporary here
 
-		A, B = diagm(1 ./(1:M)/4), zeros(M, M)
-
-		c, d = hcat(zeros(M, 1), vcat(1/4, zeros(M-1, 1))), reshape((-1).^(0:M-1), M, 1)
+		A = diagm(1 ./(1:M)/4)
+		B = zeros(M, M)
+		c = hcat(zeros(M, 1), vcat(1/4, zeros(M-1, 1)))
+		d = reshape((-1).^(0:M-1), M, 1)
 
 		# Set values of B depending on cases
 		
@@ -86,36 +87,58 @@ function BuildLinSys(M::Int, λ::Union{Vector,Number}, μ::Union{Vector,Number};
 			
 			B₀(j, k) = 4*(-1)^(j-k+1)/((2j-2k-1)*(2j-2k+1)*(2j+2k+3)*(2j+2k+5))/π
 
-			[[B[j+1, k+1] = B₀(j, k) for j = 0:M-1] for k = 0:M-1]
+			for j = 0:M-1
+
+				for k = 0:M-1
+
+					B[j+1, k+1] = B₀(j, k)
+
+				end
+			end
 			
 		else
 
 			# When analytic solution does not exists, use numerical integration
 			
-			D_func(ξ) = @. sqrt(ξ^2 + μ) * tanh(sqrt(ξ^2 + μ) / λ[1]) + λ[2]
+			D_func(ξ) = @. sqrt(ξ^2 + μ) * tanh(sqrt(ξ^2 + μ) / λ[1]) + λ[2]	# SQG kernel function
 
-			[[B[j+1, k+1] = JJ_int(x -> 1 ./(D_func(x).*x), j, k, tol)[1] for j = 0:M-1] for k = 0:M-1]
+			for j = 0:M-1
+
+				for k = 0:M-1
+
+					B[j+1, k+1] = JJ_int(x -> 1 ./(D_func(x).*x), j, k, tol)[1]
+				end
+			end
 			
 		end
 
 	else
-
-		# Set values of A, B, c, d for LQG case
+		
+		# Get number of layers from input size
 
 		N = length(μ)
 
-		A, B₀ = zeros(N*M, N*M), zeros(N*M, N*M)
+		# Set temporary values for A, B, c, d for LQG case
+
+		A  = zeros(N*M, N*M)
+		B₀ = zeros(N*M, N*M) 
+		B  = zeros(N*M, N*M, N)
+		c  = zeros(N*M, N+1)
+		d  = zeros(N*M, N)
+		c₀ = vcat(ones(N), zeros((M-1)*N))
 
 		# Set A and temporary B values using numerical integration
 
-		[[A[j*N.+(1:N), k*N.+(1:N)] .= JJ_int(x -> A_func(x, λ, μ), j, k, tol)[1] for j = 0:M-1] for k = 0:M-1]
+		for j = 0:M-1
 
-		[[B₀[j*N.+(1:N), k*N.+(1:N)] .= JJ_int(x -> B_func(x, λ, μ), j, k, tol)[1] for j = 0:M-1] for k = 0:M-1]
+			for k = 0:M-1
 
-		# Define temporary values for B, c, d
+				 A[j*N.+(1:N), k*N.+(1:N)] .= JJ_int(x -> A_func(x, λ, μ), j, k, tol)[1]
[…]
