### 2024-11-12 c02082a
Tidy up convoluted logic in CreateGrid to build (x,y,kr,k) in a single block, also fix TestLRD (hopefully for real)
