### 2024-11-12 a32ebe9
Replace commonly used x,y and r,θ assignment with function
