+++ docs/src/index.md
@@ -242,6 +242,8 @@ Full descriptions can be found on the [List of Functions](https://mncrowe.github
 | `Eval_b_SQG` | `src/create_modon.jl` | Function | Evaluates b at specified depths in the SQG model |
 | `Eval_w_SQG` | `src/create_modon.jl` | Function | Evaluates w at (non-surface) depths in the SQG model |
 | `Calc_∇` | `src/create_modon.jl` | Function | Calculates the x and y derivatives of a given field |
+| `CartesianGrid` | `src/create_modon.jl` | Function | Outputs 2D (x, y) Arrays from `grid` |
+| `PolarGrid` | `src/create_modon.jl` | Function | Outputs 2D (r, θ) Arrays from `grid` |
 
 
 [^1]: [Johnson, E. R., and M. N. Crowe, 2023, Oceanic dipoles in a surface quasigeostrophic model, J. Fluid Mech., 958, R2](https://doi.org/10.1017/jfm.2023.87).
