+++ docs/src/Functions.md
@@ -46,4 +46,6 @@ Eval_q_SQG
 Eval_b_SQG
 Eval_w_SQG
 Calc_∇
+CartesianGrid
+PolarGrid
 ```
