### 2024-11-12 9cc5251
Fix bug with PolarGrid
