### 2024-11-12 150bc1a
update ΔNCalc for consistency with A_func and B_func
