### 2024-11-12 604ffe1
Finish copying accross monoples implementation to src/monopoles.jl
