+++ src/monopoles.jl
@@ -5,9 +5,11 @@ too. These functions are included in this file.
 """
 
 """
-Function: `CreateRankine(grid, ℓ=1, x₀=[0, 0])`
+Function: `CreateRankine(grid, ℓ=1, Γ=2π, x₀=[0, 0])`
 
-Calculates the Rankine vortex for a 1 layer system
+Calculates the Rankine vortex for a 1 layer system. This vortex appears as a
+point vortex in the far field but consists of solid body rotation within the
+region r < ℓ.
 
 Arguments:
  - `grid`: grid structure containing x, y, and Krsq
@@ -16,28 +18,31 @@ Arguments:
  - `x₀`: position of vortex center, vector (default: `[0, 0]`)
 
 Note: This function outputs (u, v) directly since the solution has discontinuous velocity at
-the vortex boundary, r = ℓ.
+the vortex boundary, r = ℓ, so derivatives evaluated with Fourier transforms exhibit Gibbs
+phenomenon.
 """
 function CreateRankine(grid, ℓ::Number=1, Γ::Number=2π, x₀::Vector=[0, 0])
 
 	# Create Cartesian and polar grids
 
-	x, y = reshape(Array(grid.x), :, 1), reshape(Array(grid.y), 1, :)
-	r, θ = @. sqrt((x-x₀[1])^2 + (y-x₀[2])^2), @. atan(y-x₀[2], x-x₀[1])
+	x, y = CartesianGrid(grid)	
+	r, θ = PolarGrid(x, y, x₀)
 
 	# Calculate ψ and q using analytic result
 
-	ψ = @.   Γ * (log(r / ℓ) * (r >= ℓ) + 1/2 * (r^2 / ℓ^2 - 1) * (r < ℓ))
-	q = @.   Γ * (2 / ℓ^2) * (r < ℓ)
-	u = @. - Γ * (1 / r * (r >= ℓ) + r / ℓ^2  * (r < ℓ)) * sin(θ)
-	v = @.   Γ * (1 / r * (r >= ℓ) + r / ℓ^2  * (r < ℓ)) * cos(θ)
+	ψ = @.   Γ / (2π) * (log(r / ℓ) * (r >= ℓ) + 1/2 * (r^2 / ℓ^2 - 1) * (r < ℓ))
+	q = @.   Γ / (2π) * (2 / ℓ^2) * (r < ℓ)
+	u = @. - Γ / (2π) * (1 / r * (r >= ℓ) + r / ℓ^2  * (r < ℓ)) * sin(θ)
+	v = @.   Γ / (2π) * (1 / r * (r >= ℓ) + r / ℓ^2  * (r < ℓ)) * cos(θ)
 
 	# Move result to GPU if `cuda=true` in grid
 
 	if grid.Krsq isa CuArray
 		
-		ψ, q = CuArray(ψ), CuArray(q)
-		u, v = CuArray(u), CuArray(v)
+		ψ = CuArray(ψ)
+		q = CuArray(q)
+		u = CuArray(u)
+		v = CuArray(v)
 		
 	end
 	
@@ -45,5 +50,129 @@ function CreateRankine(grid, ℓ::Number=1, Γ::Number=2π, x₀::Vector=[0, 0])
 
 end
 
+""
+Function: `CreateMonopole(grid, ℓ=1, Γ=2π, R=Inf, x₀=[0, 0])`
 
+Calculates a monopolar vortex satisfying a Long's model assumption q = F(ψ)
+where q = [∇²-1/R²]ψ. We take F(z) = -(K²+1/R²)z for r < ℓ and F(z) = 0 for
+r > ℓ. These solutions exist only on an f-plane (β = 0).
 
+Arguments:
+ - `grid`: grid structure containing x, y, and Krsq
[…]
