### 2024-11-12 f6da9d4
Fix missing quotation mark in docstring
