### 2024-11-12 6919dbe
Fix bug with CreateMonopole and add Base.show and Base.summary functions for GridStruct
