### 2024-11-12 9fecbba
Add Base.show and Base.summary to documentation
