### 2024-11-14 2878d3e
include Energy and Enstrophy functions
