+++ src/vortex_types.jl
@@ -1,0 +1,35 @@
+"""
+This file contains functions which build vortices as a `VortexSolution` type.
+
+
+Plan:
+
+Single type, `VortexSolution` which can be SQG or LQG (maybe Monopole?)
+
+show and summary functions to display properties
+
+Maybe types for SQG and LQG parameter sets?
+
+
+
+"""
+
+
+"""
+Function: `...`
+
+...
+
+Arguments:
+ - `n`: order, Integer
+ - `x`: evaluation point, Number or Array
+
+Note: ...
+"""
+function ...
+	
+	
+
+	return ...
+end
+
