### 2024-11-14 668a1c1
add tests for energetics.jl functions
