### 2024-11-14 d6c5883
add examples for energetics functions
