### 2024-11-15 d0f6a2c
Fix recurring issue where Nothing was used instead of nothing; incorporate structs for vortices and parameter sets
