+++ src/vortex_types.jl
@@ -1,35 +1,384 @@
 """
-This file contains functions which build vortices as a `VortexSolution` type.
+This file contains functions which build vortices and vortex parameter sets as structures.
 
-
-Plan:
-
-Single type, `VortexSolution` which can be SQG or LQG (maybe Monopole?)
-
-show and summary functions to display properties
-
-Maybe types for SQG and LQG parameter sets?
-
-
+These functions are intended to allow all fields and variables associated with a given
+solution to be included in the same structure and allow parameters to be checked.
 
 """
 
-
 """
-Function: `...`
+Structure: `LQGParams`
 
-...
+Stores the parameters for an LQG dipolar vortex solution
 
 Arguments:
- - `n`: order, Integer
- - `x`: evaluation point, Number or Array
-
-Note: ...
+ - `U`: Vortex speed
+ - `ℓ`: Vortex radius
+ - `R`: Rossby radius
+ - `β`: background PV gradient
+ - `ActiveLayers`: 1 => layer contains vortex region
+ - `H`: thickness of each layer
+ - `x₀`: Vortex position
+ - `α`: Direction of vortex propagation
+ - `M`: number of coefficients in Zernike expansion
+ - `tol`: maximum error in solution evaluation
+ - `K₀`: initial guess for eigenvalue
+ - `a₀`: initial guess for coefficients
+ - `use_analytic`: use analytic solution (1-layer only)
+ - `CalcVelocity`: flag to determine if velocity is calculated
+ - `CalcEnergy`: flag to determine if energy is calculated
+ - `CalcEnstrophy`: flag to determine if enstrophy is calculated
 """
-function ...
-	
-	
-
-	return ...
+struct LQGParams
+	U::Number
+	ℓ::Number
+	R::Union{Number,Vector}
+	β::Union{Number,Vector}
+	ActiveLayers::Union{Number,Vector}
+	H::Union{Number,Vector}
+	x₀::Vector
+	α::Number
+	M::Int
+	tol::Number
+	K₀::Union{Number,Array,Nothing}
+	a₀::Union{Array,Nothing}
+	use_analytic::Bool
+	CalcVelocity::Bool
+	CalcEnergy::Bool
+	CalcEnstrophy::Bool
 end
 
+"""
+Structure: `SQGParams`
+
+Stores the parameters for an SQG dipolar vortex solution
+
+Arguments:
+ - `U`: Vortex speed
+ - `ℓ`: Vortex radius
+ - `R`: Rossby radius
+ - `β`: background PV gradient
+ - `x₀`: Vortex position
+ - `α`: Direction of vortex propagation
+ - `M`: number of coefficients in Zernike expansion
+ - `tol`: maximum error in solution evaluation
+ - `K₀`: initial guess for eigenvalue
+ - `a₀`: initial guess for coefficients
+ - `CalcVelocity`: flag to determine if velocity is calculated
+ - `CalcEnergy`: flag to determine if energy is calculated
+ 
+"""
+struct SQGParams
+	U::Number
+	ℓ::Number
+	R::Vector
+	β::Number
+	x₀::Vector
+	α::Number
+	M::Int
+	tol::Number
+	K₀::Union{Number,Array,Nothing}
+	a₀::Union{Array,Nothing}
+	CalcVelocity::Bool
+	CalcEnergy::Bool
+end
+
+"""
+Structure: `LQGVortex`
+
[…]
