+++ src/vortex_types.jl
@@ -24,7 +24,7 @@ Arguments:
  - `tol`: maximum error in solution evaluation
  - `K₀`: initial guess for eigenvalue
  - `a₀`: initial guess for coefficients
- - `use_analytic`: use analytic solution (1-layer only)
+ - `UseAnalytic`: use analytic solution (1-layer only)
  - `CalcVelocity`: flag to determine if velocity is calculated
  - `CalcEnergy`: flag to determine if energy is calculated
  - `CalcEnstrophy`: flag to determine if enstrophy is calculated
@@ -42,7 +42,7 @@ struct LQGParams
 	tol::Number
 	K₀::Union{Number,Array,Nothing}
 	a₀::Union{Array,Nothing}
-	use_analytic::Bool
+	UseAnalytic::Bool
 	CalcVelocity::Bool
 	CalcEnergy::Bool
 	CalcEnstrophy::Bool
@@ -104,8 +104,8 @@ struct LQGVortex
 	params::LQGParams
 	ψ::Union{CuArray,Array}
 	q::Union{CuArray,Array}
-	K::Union{Array,Nothing}
-	a::Array
+	K::Array
+	a::Union{Array,Nothing}
 	u::Union{CuArray,Array,Nothing}
 	v::Union{CuArray,Array,Nothing}
 	KE::Union{Vector,Nothing}
@@ -130,7 +130,7 @@ Arguments:
  - `SPE`: surface potential energy
 """
 struct SQGVortex
-	params::LQGParams
+	params::SQGParams
 	ψ::Union{CuArray,Array}
 	b::Union{CuArray,Array}
 	K::Array
@@ -159,7 +159,7 @@ Arguments:
  - `tol`: maximum error in solution evaluation
  - `K₀`: initial guess for eigenvalue
  - `a₀`: initial guess for coefficients
- - `use_analytic`: use analytic solution (1-layer only)
+ - `UseAnalytic`: use analytic solution (1-layer only)
  - `CalcVelocity`: flag to determine if velocity is calculated
  - `CalcEnergy`: flag to determine if energy is calculated
  - `CalcEnstrophy`: flag to determine if enstrophy is calculated
@@ -176,13 +176,13 @@ function DefLQGParams(U::Number=1,
 	            tol::Number=1e-6,
 		     K₀::Union{Number,Array,Nothing}=nothing,
 		     a₀::Union{Array,Nothing}=nothing,
-	   use_analytic::Bool=false,
+	    UseAnalytic::Bool=false,
 	   CalcVelocity::Bool=false,
              CalcEnergy::Bool=false,
 	  CalcEnstrophy::Bool=false)
 	
 	return LQGParams(U, ℓ, R, β, ActiveLayers, H, x₀, α, M,
-		tol, K₀, a₀, use_analytic, CalcVelocity, CalcEnergy, CalcEnstrophy)
+		tol, K₀, a₀, UseAnalytic, CalcVelocity, CalcEnergy, CalcEnstrophy)
 end
 
 """
@@ -206,7 +206,7 @@ Arguments:
 """
 function DefSQGParams(U::Number=1,
 		      ℓ::Number=1,
-		      R::Vector=Inf,
+		      R::Vector=[Inf, Inf],
 		      β::Number=0,
 		     x₀::Vector=[0, 0],
 		      α::Number=0,
@@ -217,7 +217,7 @@ function DefSQGParams(U::Number=1,
 	   CalcVelocity::Bool=false,
              CalcEnergy::Bool=false)
 
-	return LQGParams(U, ℓ, R, β, x₀, α, M, tol, K₀, a₀, CalcVelocity, CalcEnergy)
[…]
