### 2024-11-15 056b04f
Move Base.show and Base.summary to seperate section on Functions.md
