### 2024-11-20 e0c730a
implement new monopoles functions including N-layer PV inversion
