+++ src/QGDipoles.jl
@@ -71,8 +71,10 @@ export
 
 	# monopoles.jl
 	CreateRankine,
-	CreateMonopole,
-	InvertVorticity1LQG
+	Create1LMonopole,
+	InvertVorticity1LQG,
+	CreateLQGMonopole,
+	InvertVorticityLQG
 
 # Include all function and type definitions
 
