### 2024-11-25 5af3795
added function to calculate vorticity for a given solution
