### 2024-11-25 d66072a
remove unnecessary list of functions in index.md, autogenerated list in functions.md is preferred
