+++ docs/src/index.md
@@ -210,47 +210,6 @@ Note: this package returns ``b/N`` rather than ``b``. When working with dimensio
 | ``f`` | Coriolis parameters | - |
 | ``H`` | layer depth | - |
 
-## Index of Functions
-
-This list contains all Functions and Structures from the QGDipoles.jl package.
-Full descriptions can be found on the [List of Functions](https://mncrowe.github.io/QGDipoles.jl/dev/Functions/) page.
-
-| Name | Location | Type | Description |
-| :--| :-- | :-- | :-- |
-| `A_func` | `src/JJ_integ.jl` | Function | Evaluates a function required to calculate the matrix ``\textbf{A}`` in the LQG case |
-| `B_func` | `src/JJ_integ.jl` | Function | Evaluates a function required to calculate the matrix ``\textbf{B}`` in the LQG case |
-| `JJ_int` | `src/JJ_integ.jl` | Function | Calculates a double Bessel function integral required to calculate ``\textbf{A}`` and ``\textbf{B}`` |
-| `BuildLinSys` | `src/lin_sys.jl` | Function | Builds the terms in the inhomogeneous eigenvalue problem; ``\textbf{A}``, ``\textbf{B}``, ``\textbf{c}`` and ``\textbf{d}`` |
-| `ApplyPassiveLayers` | `src/lin_sys.jl` | Function | Removes rows and columns corresponding to passive layers from the linear system |
-|`IncludePassiveLayers` | `src/lin_sys.jl` | Function | Includes columns corresponding to passive layers in the eigenvalue and coefficient arrays |
-| `SolveInhomEVP` | `src/lin_sys.jl` | Function | Solves the inhomogeneous eigenvalue problem using nonlinear root finding (``N > 1``) or standard eigenvalue methods for quadratic problems (``N = 1``) |
-| `InhomEVP_F!` | `src/lin_sys.jl` | Function | Calculates the function required for `SolveInhomEVP` and it's derivatives |
-| `OrthogSpace ` | `src/lin_sys.jl` | Function | Extends the input to an orthonormal basis over ``R^n`` using the Gram-Schmidt method, required for `SolveInhomEVP` |
-| `ZernikeR` | `src/create_modon.jl` | Function | Define the Zernike radial function using the `jacobi` function from `SpecialFunctions.jl` |
-| `GridStruct` | `src/create_modon.jl` | Structure | A structure that stores the grid variables in physical and Fourier space |
-| `CreateGrid` | `src/create_modon.jl` | Function | Define the numerical grid in the form of a `GridStruct` structure |
-| `Calc_ψq` | `src/create_modon.jl` | Function | Calculate ``\psi`` and ``q`` in a layered QG model using coefficients and vortex parameters |
-| `Calc_ψb` | `src/create_modon.jl` | Function | Calculate ``\psi`` and ``b`` in the SQG model using coefficients and vortex parameters |
-| `Calc_uv` | `src/create_modon.jl` | Function | Calculate the velocity fields from ``\psi`` using ``(u, v) = (-\partial\psi/\partial y, \partial\psi/\partial x)`` |
-| `ΔNCalc` | `src/create_modon.jl` | Function | Defines a matrix used to invert for ``\psi`` and ``q`` in Fourier space |
-| `CreateModonLQG` | `src/create_modon.jl` | Function | High level wrapper function for calculating ``\psi``, ``q``, ``K`` and ``\textbf{a}`` for the Layered QG model using given parameters |
[…]
