### 2024-12-10 #7448 @kthyng
@mncrowe Can you comment on the size changes of the repository? What has changed in the repository?

Several editors have the question of whether this should be a stand alone package or whether we should be encouraging you to make this package a contribution to another more general package (ex. GeophysicalFlows.jl). Why are you creating a separate package for this work? How could it be expanded if it is going to stand alone?
