### 2024-12-10 #7448 @mncrowe
Regarding the changes to the repository; this was to implement some additional features:

- Custom types for parameter sets and vortex solutions
- Energy, Enstrophy and vorticity diagnostic functions so these quantities do not have to be evaluated manually (or by using another package)
- Some lines that were used multiple times (e.g. defining polar coordinates) have been move to their own functions
- Added some monopolar vortex solutions

My reason for keeping it as a separate package is essentially that merging it would reduce it's reach and applicability:

- QGDipoles can be used to set up initial conditions for various packages and I don't want to tie it to any one. It was made to be compatible with GeophysicalFlows as that was the package I was using most at the time, however it will also work with other packages too (potentially with some reshaping of arrays). Tying it to just one package would likely prevent it's use by users of other packages.
- QGDipoles currently contains functionality that these general PDE solver packages do not. For example, my SQG solutions allow for finite depth whereas the SQG model in GeophysicalFlows only allows for an infinite depth. As such, there would be various consistency issues when merging this package with others. These issues may force me to reduce or limit the functionality I'm offering and restrict future development.
- The solutions calculated by this package aren't intended solely to be used as initial conditions for a initial value problem (IVP) so packaging QGDipoles with an IVP solver restricts it's use. For example, there are people working on theoretical vortex dynamics who would not want to run time-dependent simulations, and only care about the form of these solutions. They would not necessarily know that an IVP solver contained the functions they were looking for and would not make use of most of the functionality offered by a combined package.
- Users of general packages are working on a wide range of problems and in many cases will not need the functionality of QGDipoles. Therefore it seems like adding QGDipoles to any of them would add unnecessary bloat that many users would not benefit from.

There are various directions for expansion, some of which I'm currently working on:

- Additional functions for calculating monopolar vortices (in progress)
- Functions for higher order poles (quadrupoles etc.)
- A 3D quasi-geostrophic model with corresponding types (in progress)
- Dipolar solutions to the shallow water equations (SWE) and magnetohydrodynamical equations (MHD)
- Functions to enable compatibility with other packages (e.g. Oceananigans, eigenvalue solver for stability problems)
- Theoretical energy predictions (in progress)
- Additional diagnostics (e.g. momentum/impulse)

These additions would cause QGDipoles to further diverge from GeophysicalFlows and many could not be added to a more general package if it did not support the underlying model (GeophysicalFlows does not support the 3D QG model, SWE or MHD and likely never will).
