### 2024-12-10 #7448 @kthyng
@mncrowe Thanks for your thorough response, this is helpful. It does bring up an additional question then: given your active development, would it be better to wait until some of your in progress work is finished to undergo JOSS review, so those features can be included in the review?
