### 2024-12-10 #7448 @mncrowe
@kthyng That is a good question, currently the three areas I'm working on (the ones listed as in progress above) are limited by the theoretical side rather than the coding side so it's difficult to put a timeframe on finishing their implementation as I'm not sure how easily the maths will work out. Once I've got the theory working, these additions will follow the existing format. For example, 3D QG will use similar types and functions to SQG and LQG, just with different calculations under the hood.

As such, I'd rather a thorough review of the codebase sooner rather than later though it is of course up to you and the editors. Any improvements suggested by the reviewers would also benefit these later implementations.
