### 2025-02-03 #7448 @rwegener2
👋🏻 @ali-ramadhan and @haakon-e, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
