### 2025-02-03 #7448 @haakon-e
Hi @rwegener2, I'd be happy to review this submission.

I would note that I have not thought deeply about vortex dynamics in a few years, so it may be useful if another reviewer is a specialist within this subfield. That said, I can certainly provide feedback on the software from the perspective of an atmosphere/ocean modeler, an applied mathematician, and a Julia developer. I would likely be able to start reviewing sometime next week. Let me know if you'd like to proceed.
