### 2025-02-10 #7767 @editorialbot
[REVIEW]: QGDipoles.jl: A Julia package for calculating dipolar vortex solutions to the Quasi-Geostrophic equations
**Submitting author:** @mncrowe (<a href="http://orcid.org/0000-0002-9916-2653">Matthew Crowe</a>)
**Repository:** https://github.com/mncrowe/QGDipoles.jl/
**Branch with paper.md** (empty if default branch): joss_draft
**Version:** v1.0.1
