### 2025-02-17 mncrowe/QGDipoles.jl#16 @haakon-e
Improve installation instructions
If I currently paste the installation instructions[^1] into the Julia REPL, I get an error:

I recommend updating the instructions to:

``````
## Installation

QGDipoles.jl is a Julia package that requires Julia version 1.10 or later ([download Julia here](https://julialang.org/downloads/)).
Launch Julia and install QGDipoles.jl by typing these commands
```julia-repl
julia> import Pkg

julia> Pkg.add(url="https://github.com/mncrowe/QGDipoles.jl")
```

``````

I recommend also updating the corresponding instructions in the [docs](https://mncrowe.github.io/QGDipoles.jl/dev/Installation/).

---
ref: https://github.com/openjournals/joss-reviews/issues/7767

[^1]: from the readme file in: https://github.com/mncrowe/QGDipoles.jl
