### 2025-02-18 mncrowe/QGDipoles.jl#16 @mncrowe
I see the issue. Copying `]` does not seem to open the package manager and it only works when typed into the REPL. I'd been under the assumption that reader would type `]` then copy the other two lines. Perhaps something like the following is a middle ground:

``````
text ... open the Julia package manager by entering `]` at the REPL and then enter the following:
```julia
add https://github.com/mncrowe/QGDipoles.jl.git
instantiate
```
``````

Could go with the `pkg.add` option but I just find using the package manager directly to be 'neater' somehow.
