### 2025-02-18 mncrowe/QGDipoles.jl#17 @haakon-e
Add developer documentation
To attract new developers to the package, I would recommend the following:

1. Add instructions for how to contribute.
2. Add instructions how to build docs
3. Add instructions how to run tests

---

# 1. Contribution instructions

This can take many forms, including:
- Adding a page in the docs (inspiration: [JuMP.jl](https://jump.dev/JuMP.jl/dev/developers/contributing/), [Oceananigans.jl](https://clima.github.io/OceananigansDocumentation/stable/contributing/), [SciML](https://docs.sciml.ai/DiffEqDevDocs/latest/))
- Adding a ColPrac badge to the `README.md` file ([link](https://docs.sciml.ai/ColPrac/stable/#Marking-a-Repository-as-following-ColPrac:))
- Adding a note in the `README.md` file

# 2. Building docs
I see that the GitHub workflow instantiates the `docs` environment by
1. activating `julia --project=docs/`
2. Running `Pkg.develop(".")` to add `QGDipoles.jl` as a new dependency to the `docs` environment.
3. Building docs by running `make.jl`

I would recommend either:
- Add instructions for this in a developer documentation section of the docs
- Add `QGDipoles` to the docs environment. I think you should be able to do something like this:
```toml
# docs/Project.toml

[deps]
Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
QGDipoles = "192f05a4-068c-458e-9e9e-a50cd96b7215"

[compat]
Documenter = "1.7"

[sources]
QGDipoles = {path = ".."}
```
The `[sources]` section is unfortunately ignored in Julia v1.10. So that `] instantiate` does not work directly:
```julia-repl
(docs) pkg> st
Status `~/Documents/reviews/QGDipoles.jl/docs/Project.toml`
  [e30172f5] Documenter
  [91a5bcdd] Plots
  [192f05a4] QGDipoles

(docs) pkg> instantiate
    Updating registry at `~/.julia/registries/General.toml`
ERROR: expected package `QGDipoles [192f05a4]` to be registered
```
For julia v1.10, you'd have to explicitly `dev` the package (here, assuming you're in `QGDipoles.jl` root directory):
```julia-repl
(docs) pkg> dev .
   Resolving package versions...
  Downloaded artifact: LLVMExtra
    Updating `~/Documents/reviews/QGDipoles.jl/docs/Project.toml`
  [e30172f5] + Documenter v1.8.1
  [91a5bcdd] + Plots v1.40.9
  [192f05a4] + QGDipoles v1.0.0 `..`

[...]
```

Alongside instructions to run `docs/make.jl` to build docs, I would also recommend an explicit note of where you find the built docs (e.g. `navigate to "docs/build/" and open (e.g.) `index.html` with your favorite web browser`)

# 3. Running tests
It should be pretty straightforward to explain how to do this. The tests ran nicely on my computer (with `include("runtests.jl")`).

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
