### 2025-02-18 mncrowe/QGDipoles.jl#18 @haakon-e
tip for `@example` blocks in docs
Just a little tip: You can hide lines of code in `@example` blocks in the docs. Currently many of the code blocks look like this:

![Image](https://github.com/user-attachments/assets/838ebd7a-eebb-48b7-896d-3cee1f4e8640)

If you add `# hide` after `nothing`, you still suppress output, but don't explicitly display the `nothing` statement in the docs, e.g.:
``````

```@example 1layer
[...]
# Set grid parameters

Nx, Ny = 512, 512
Lx, Ly = 10, 10

nothing # hide
```

``````
now your `@example` blocks will look a bit cleaner :-)

![Image](https://github.com/user-attachments/assets/025762bc-6c1e-42ed-8ea0-f854d0c96d77)

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
