### 2025-02-18 mncrowe/QGDipoles.jl#16 @haakon-e
I totally agree that experienced developers will prefer to type `]` then copy the other lines. And they will probably do that based solely on you stating whether the package is registered or not (they will know to copy the URL from their web browser).

This suggestion is explicitly geared towards minimizing barriers for users who are not experienced with Julia.
