### 2025-02-18 mncrowe/QGDipoles.jl#19 @haakon-e
Add a `Project.toml` in the `examples` folder
Running the example scripts, or following the `Examples` section in the docs, requires additional packages. Notably, `Plots` and `GeophysicalFlows`[^1].

To simplify the process for users[^2], I would recommend either:

## Create `examples/Project.toml` file
Create a `examples/Project.toml` file is created with the content:
```toml
# examples/Project.toml

[deps]
GeophysicalFlows = "44ee3b1c-bc02-53fa-8355-8e347616e15e"
Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
QGDipoles = "192f05a4-068c-458e-9e9e-a50cd96b7215"
```

then, somewhere in the `Examples` section of the docs, instruct users to activate that environment to follow along easily.

A downside with this suggestion is that it requires users to download the package source code. Neither the current, nor suggested[^3], installation instructions tell users they need to clone the source code.

## Update `Examples` docs
If you keep installation instructions as-is (`add`ing instead of `git clone`ing), I would recommend:
1. Tell users how they can create a new environment in which they install `QGDipoles.jl`.
    - This avoids inadvertently polluting inexperienced Julia users' base environment `(@v1.11)` with many packages.
2. Tell users to add `Plots` (and when applicable, `GeophysicalFlows`) to that environment, so that they can run the examples locally.

[^1]: Since only one example requires `GeophysicalFlows`, that example could have its own `Project.toml` file. But then I think it would need to live in its own (sub)folder.

[^2]: this is really geared towards inexperienced Julia users. Experienced users will very likely understand how to install `Plots.jl` themselves.

[^3]: Ref suggested update to installation instructions: https://github.com/mncrowe/QGDipoles.jl/issues/16

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
