### 2025-02-19 mncrowe/QGDipoles.jl#20 @haakon-e
Add unit tests for core methods/structs
It would be great to see simple unit tests that verifies basic functionality for core structs and methods.

For example, the `CreateGrid` function appear to be untested in CI for vector inputs of `Lx` and `Ly`[^1].

https://github.com/mncrowe/QGDipoles.jl/blob/d66072a268e2320f66e9363995ec9e3666b02a2f/src/create_modon.jl#L95-L105

[^1]: The [`3_Layer_Modon.jl`](https://github.com/mncrowe/QGDipoles.jl/blob/d66072a268e2320f66e9363995ec9e3666b02a2f/examples/3_Layer_Modon.jl#L17-L35) example (`Example 2` in the docs) does demonstrate usage of vector `Lx` and `Ly`

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
