### 2025-02-19 mncrowe/QGDipoles.jl#21 @haakon-e
Unexpected behavior for `CreateGrid`
If I understand the docstring for `CreateGrid` correctly,
https://github.com/mncrowe/QGDipoles.jl/blob/d66072a268e2320f66e9363995ec9e3666b02a2f/src/create_modon.jl#L95-L105

if I supply `Lx`/`Ly` as vectors, they should have two elements, with each element indicating the left- and right endpoints. So `Example 1` in the docs,
```julia
# Set grid parameters

Nx, Ny = 512, 512
Lx, Ly = 10, 10

# [...]

grid = CreateGrid(Nx, Ny, Lx, Ly)
GridStruct
  ├────────────────────── device: CPU
  ├─────────────── size (Lx, Ly): (10.0, 10.0)
  ├───────── resolution (Nx, Ny): (512, 512)
  ├─────── grid spacing (Δx, Δy): (0.01953125, 0.01953125)
  └────────────────────── domain: x ∈ [-5.0, 4.98046875]
                                  y ∈ [-5.0, 4.98046875]
```
should be equivalently written as
```julia
Lx, Ly = [-5, 5], [-5, 5]

grid = CreateGrid(Nx, Ny, Lx, Ly)
GridStruct
  ├────────────────────── device: CPU
  ├─────────────── size (Lx, Ly): (5.0, 5.0)
  ├───────── resolution (Nx, Ny): (512, 512)
  ├─────── grid spacing (Δx, Δy): (0.009765625, 0.009765625)
  └────────────────────── domain: x ∈ [-5.0, -0.009765625]
                                  y ∈ [-5.0, -0.009765625]
```
however, as you can see, this results in a domain from -5 to 0, instead of -5 to 5.

If it's the case that I misinterpreted the docstring, I would recommend updating it. Otherwise, a bug fix is needed.

cc: https://github.com/mncrowe/QGDipoles.jl/issues/20

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
