### 2025-02-19 mncrowe/QGDipoles.jl#22 @haakon-e
Simplify plotting by creating a `@recipe`
Your preferred plotting package, `Plots.jl`, provides functionality simplifying visualization of custom user types using [recipes](https://docs.juliaplots.org/stable/recipes/)

You can simplify plotting for all examples in the docs with the following suggested recipe:
```julia
@recipe function f(grid::GridStruct, ψ::AbstractArray)
	# Set default attributes
	aspect_ratio --> 1
	colormap --> :balance
	xlims --> (minimum(grid.x), maximum(grid.x) + step(grid.x))
	ylims --> (minimum(grid.y), maximum(grid.y) + step(grid.y))
	xlabel --> "x"
	ylabel --> "y"

	# Return the plotted values
	grid.x, grid.y, transpose(ψ)
end
```
which lets you simply type
```julia
heatmap(grid, ψ[:,:,1])  # Example 2 solution
```
to generate a figure

You can also consider adding a custom keyword to select layers:
```julia
@recipe function f(grid::GridStruct, ψ::AbstractArray; layer = 1)
	# Set default attributes
	aspect_ratio --> 1
	colormap --> :balance
	xlims --> (minimum(grid.x), maximum(grid.x) + step(grid.x))
	ylims --> (minimum(grid.y), maximum(grid.y) + step(grid.y))
	xlabel --> "x"
	ylabel --> "y"

	# Return the plotted values
	grid.x, grid.y, transpose(@view(ψ[:, :, layer]))
end
```
then
```julia
heatmap(grid, ψ, layer=2)
```
gives

With a definition like `@recipe function f(grid::GridStruct, ψ::CuArray; layer = 1)` you can also abstract away copying memory from GPU to CPU (ref: [Example 5](https://mncrowe.github.io/QGDipoles.jl/dev/Examples/#Example-5:-A-GeophysicalFlows.jl-simulation)).

To achieve this in practice, you'd need to either
- add `RecipesBase` as a direct dependency (it's in fact already an indirect dependency due to `QGDipoles → Jacobi → Polynomials → RecipesBase`)
- write the recipe in an extension ([official docs](https://docs.julialang.org/en/v1/manual/code-loading/#man-extensions), [implementation example](https://github.com/pebeto/julia_extensions_example)), so that only users who prefer `Plots.jl` need to care about the visualization code.

Since `QGDipoles.jl` already depends on `RecipesBase` through `Jacobi`, I'd recommend going with the first approach.

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
