### 2025-02-20 mncrowe/QGDipoles.jl#23 @haakon-e
`docs/readme.md`
Given the existing `README.md` in root as well as the online documentation, what is the purpose of the `docs/readme.md` file?

It seems to me like just more text to keep up to date as the package evolves.

I'd recommend removing it and rather point users to the built docs.

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
