### 2025-02-20 mncrowe/QGDipoles.jl#24 @haakon-e
Refactor user-facing API
I think usage would be less error-prone if user-facing functions are primarily keyword-based. Internal functions can remain as-is.

This change would also facilitate usage the provided default values more easily.

## For example, with `CreateModonLQG`
https://github.com/mncrowe/QGDipoles.jl/blob/d66072a268e2320f66e9363995ec9e3666b02a2f/src/create_modon.jl#L454-L456
it is currently not possible to specify a problem setup without also specifying `M`. However, in `Example 1` in the docs, it is stated that

> We've taken ``M = 8`` as this is generally a sufficient number of terms to get a relative error ``< 10^{-6}`` in the final result.

if this is indeed true, it seems it would be quite useful to simply leave that value at the default of 8. Currently this is not possible, since it is a positional argument.

I suggest defining `CreateModonLQG` e.g. as:
```julia
function CreateModonLQG(;
	grid,
	M::Int = 8,
	U::Number = 1,
	ℓ::Number = 1,
	R::Union{Number,Vector} = 1,
	β::Union{Number,Vector} = 0,
	ActiveLayers::Union{Number,Vector} = 1,
	x₀::Vector = [0, 0],
	α::Number = 0,
	K₀::Union{Number,Array,Nothing} = nothing,
	a₀::Union{Array,Nothing} = nothing,
	tol = 1e-6,
)
```
so its usage in [`Example 4`](https://mncrowe.github.io/QGDipoles.jl/dev/Examples/#Example-4:-Wrappers) could simply be:
```julia
# create modon solution

ψ, q, K, a = CreateModonLQG(; grid, U, ℓ, R, β, x₀)
# no need for `M`, `tol`, `ActiveLayers`
```

## Argument order inconsistencies

Notwithstanding the changes above, there seems to be some inconsistencies between different user-facing methods.

For example `CreateModonLQG` and `DefLQGVortex` orders its arguments (notably `M`) differently. Please review to ensure consistency.

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
