### 2025-02-20 mncrowe/QGDipoles.jl#25 @haakon-e
Tip for rendering of functions in docs
A minor note / tip: `Documenter.jl` automatically recognizes the first line of a docstring as the function definition(s) ([ref](https://documenter.juliadocs.org/stable/man/guide/#Adding-Some-Docstrings)), and can render math equations.

For example, if [this function](https://github.com/mncrowe/QGDipoles.jl/blob/d66072a268e2320f66e9363995ec9e3666b02a2f/src/JJ_integ.jl#L23-L33)

was documented like this instead:
```julia
"""
	A_func(ξ, λ, μ)

Evaluates the matrix function ``A(ξ, λ, μ) = K(ξ) [K(ξ) + D(μ)]⁻¹ ξ⁻¹``

# Arguments:
 - `ξ`: point in ``[0, ∞)``, `Number`
 - `λ`: ratio of vortex radius to Rossby radius in each layer, `Number` or `Vector`
 - `μ`: nondimensional (y) vorticity gradient in each layer, `Number` or `Vector`
"""
```
it would look like this:

I would recommend reviewing all docstrings and update them following the example above.

Key points:
- Indent the function name, removing the `Function:` prefix
- Prefix `Arguments:` with `#` (i.e. `# Arguments:`), which makes it a header
- Use single backticks (`) to render types and argument names as "code" (e.g. `Number`)
- Use double backticks (``) around math so it gets rendered nicely.
    - For centered math, make a math environment with:
``````
```math
ax=b
```
``````

----

For other math, like
https://github.com/mncrowe/QGDipoles.jl/blob/d66072a268e2320f66e9363995ec9e3666b02a2f/src/JJ_integ.jl#L108-L110
it already renders nicely online

but the docstring can be made more readable. Instead of
```julia
"""
``I = \\int_0^\\infty F(\\xi) J_{2j+2}(\\xi) J_{2k+2}(\\xi) \\mathrm{d}\\xi``
"""
```
you can display the unicode directly:
```julia
"""
``I = ∫_0^∞ F(ξ) J_{2j+2}(ξ) J_{2k+2}(ξ) \\mathrm{d}ξ``
"""
```
which still displays nicely online

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
