### 2025-02-20 mncrowe/QGDipoles.jl#26 @haakon-e
`List of Functions` doc page
# Ordering/grouping of functions
I think the API reference would be more useful if it grouped "user facing" (or "public") methods and "internal" (or "private") methods in separate sections.

As it stands now, the first method I see is `QGDipoles.A_func`

and as far as I can tell, a user should never call this method directly. I think it would be better to clearly label functions like this as "internal".

Another benefit of labeling them as "internal" is that if you ever need to change them in the future, you don't have to make a breaking release.

related to: https://github.com/mncrowe/QGDipoles.jl/issues/24

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
