### 2025-02-20 7dc41aa
refactor: remove unused vars
