### 2025-02-21 mncrowe/QGDipoles.jl#28 @haakon-e
Improve support for new users
At a high level, I would recommend thinking about ways to improve the experience for new users[^1].

A framework I find useful for structuring the documentation is [Diátaxis](https://diataxis.fr/start-here/). In this framework, your documentation addresses needs of different types of users: learners and practitioners.
- Learners need background material (explanations) and lessons (tutorials).
- Practitioners need technical descriptions (API reference) and how-to guides

----

# `Home` section in the docs

The `About` subsection provides a technical introduction to the main functionality of the package. The text appears to be mostly suited for practitioners already familiar with much of the theory and literature in deciding whether this package can be used in their work.

On the other hand, the "Statement of need" in the JOSS paper does a great job at contextualizing the problem this package aims to solve, and is presented at a level and pace suitable for learners.

- I would recommend the `Home` page to have a very high-level overview of what `QGDipoles.jl` is.
    - The text could incorporate elements from the JOSS paper and/or the introductions of the primary cited papers[^2].
- Then, I would recommend the page to provide an overview or roadmap for the documentation[^3].
- I would move the `Method Summary` and `Equations and Parameters` subsections to a new `Background material` section.
    - It would be great if this section elaborates a bit further on the context of this package (e.g. use more elements from the JOSS paper and/or primary cited papers). Introduce the QG and then N-layer QG equations. Also add some discussion of the SQG system.
    - Then you can explain vortex solutions
    - Then provide a section that results with the linear algebra system
    - And finally the approaches to solving the linear system

# `Examples` section in the docs

Overall, I think the examples demonstrates nicely various aspects of the package. However, additional context and some restructuring may be useful.

In some cases it may be useful to simplify the examples as not to overwhelm users with knobs they may not use the first time they familiarize themselves with the package. E.g.: most knobs should have reasonable defaults.

I would recommend starting with a `quick start` section, that demonstrates the key promise of the package, simplified beyond the current `Example 1`.

With the current [`Example 1`](https://mncrowe.github.io/QGDipoles.jl/dev/Examples/#Example-1:-1-layer-QG) and [`Example 4`](https://mncrowe.github.io/QGDipoles.jl/dev/Examples/#Example-4:-Wrappers) as starting points, and with the changes proposed in #22 and #24, it could look something like:
```julia
using QGDipoles, Plots

# Set problem parameters

U, ℓ = 1, 1	# vortex speed and radius
R = 1		# Rossby radius
β = 1		# background PV gradient

# Set grid parameters

Nx, Ny = 512, 512
Lx, Ly = 10, 10

grid = CreateGrid(Nx, Ny, Lx, Ly)

# create modon solution

ψ, q, K, a = CreateModonLQG(; grid, U, ℓ, R, β)

# Visualize solution

heatmap(grid, ψ)
```

then the next example could be multi-layer QG (i.e. [`Example 2`](https://mncrowe.github.io/QGDipoles.jl/dev/Examples/#Example-2:-multi-layer-QG)), analogously simplified. That is, the only new concept introduced is how to modify the above parameters to produce multi-layer solutions:
```julia
R = [1, 1, 1]			# Rossby radius in each layer
β = [0, 0, 1]			# background PV gradient in each layer
ActiveLayers = [0, 1, 0]	# 1 => layer contains vortex region

# create modon solution

ψ, q, K, a = CreateModonLQG(; grid, U, ℓ, R, β)

# Visualize solution

heatmap(grid, ψ; layer = 2)
```

then, in the next example the only new concept introduced should be how to use SQG,
```julia
R = [Inf, Inf]	# Baroclinic and Barotropic Rossby radii
β = 1		# background PV gradient

ψ, q, K, a = CreateModonSQG(; grid, U, ℓ, R, β)

heatmap(grid, ψ)
```

I think the `GeophysicalFlows.jl` example is great! I only have a few minor comments/questions:
- I would set `dev = CPU()` since some users may not be following the examples on a device with an NVIDIA GPU.
- why do you need this line? In my testing, `q₀` is already a matrix
```julia
q₀ = reshape(q₀, nx, ny)		# convert from size (nx, ny, 1) to size (nx, ny)
```
- Could you re-upload a version of the final plot with a white background? It's currently hard to read the axis labels in browsers with dark mode enabled:

# A new `how-to` / `FAQ` section
For more advanced workflows for "practitioners", I would recommend moving explanations of many knobs to a new, more tailored section. This includes questions like:
- How do I set up the problem on a GPU?
- How do I obtain a more accurate solution?
    - introduce `M` and `tol` keywords
- How do I change the domain limits?
    - Introduce `Lx`, `Ly` as 2-vectors, e.g. `Lx = [2, 30]`
- How do I change the initial location and angle of the vortex?
    - Introduce `x₀` and `α`.
- Is there a better method to find solutions?
    - Introduce `K₀`, `a₀` and the `method = 1` flag
- How do I inspect the linear system that is being solved?
    - basically `Example 1` as it is now, with `BuildLinSys`
    - Then you can show how to obtain a solution from the linear system with `SolveInhomEVP`
- How do I calculate potential and kinetic energy for my solution?
    - Can also discuss domain-averaging
- How do I calculate enstrophy for my solution?
    - Can also discuss domain-averaging
- How do I calculate velocity for my solution?
- How do I calculate vorticity for my solution?
- How do I calculate the gradient of a field?
    - The `Calc_∇` function

The last 5 bullet points (`How do I Calculate X`) could alternatively live in a separate `Diagnostics` section of the docs.

PS: You could also frame all of the above as "How to X".

---

ref: https://github.com/openjournals/joss-reviews/issues/7767

[^1]: As I noted in the pre-review, I'm not an expert in this specific subdomain, so some of the assessments may be flawed or incomplete. If you think so, please say so.

[^2]: e.g. Crowe and Johnson [2023](https://doi.org/10.1017/jfm.2023.607), [2024](https://doi.org/10.1017/jfm.2024.619)

[^3]: Inspiration for what a roadmap could look like: [JuMP](https://jump.dev/JuMP.jl/dev/#How-the-documentation-is-structured), [Diátaxis](https://diataxis.fr/#contents)
