### 2025-02-21 mncrowe/QGDipoles.jl#29 @haakon-e
Figure 1 in JOSS paper
For completeness, it would be useful if there was a script in the `paper` folder (`joss_draft` branch) that generates Figure 1 in the JOSS paper.

Additionally, please make sure the labels and ticks have (nearly) the same font size as the paper text.

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
