### 2025-02-21 mncrowe/QGDipoles.jl#28 @haakon-e
In reference to #17 and the `Community guidelines` reviewer checkbox

> **Community guidelines**: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Please also provide sections of the docs that detail how to report issues/problems and how to seek support.
