### 2025-02-21 mncrowe/QGDipoles.jl#31 @haakon-e
Code formatting
I would recommend using a code formatter like [`JuliaFormatter.jl`](https://github.com/domluna/JuliaFormatter.jl) to provide consistent formatting of the package. This makes it easier for new contributors to contribute to the package, and for the code to have a consistent look.

Usage of the code formatter should be documented in the developer/contributor docs.

---

ref: https://github.com/openjournals/joss-reviews/issues/7767
