### 2025-02-21 #7767 @haakon-e
@rwegener2,  @mncrowe,

I've now provided feedback on most aspects of the package. I look forward to engaging directly in the issues as questions arise, and in PRs that implement resolutions.

I will put aside this review until the end of next week or so, to focus on other work.
