+++ src/JJ_integ.jl
@@ -31,34 +31,37 @@ Arguments:
  - `μ`: nondimensional (y) vorticity gradient in each layer, Number or Vector
 """
 function A_func(ξ::Number, λ::Union{Vector,Number}, μ::Union{Vector,Number})
-	
-	N = length(μ)
-	
-	if N == 1
 
-		# Calculate A in 1-layer case
-		
-		K = @. ξ^2 + λ^2
+    N = length(μ)
 
-		A = @. K / (K + μ) / ξ
+    if N == 1
 
-	else
+        # Calculate A in 1-layer case
 
-		# Calculate A in N-layer case (N > 1)
-		
-		diagonal_elements       = [λ[1]^2; 2*λ[2:end-1].^2; λ[end]^2]
-		above_diagonal_elements = -λ[1:end-1].^2
-		below_diagonal_elements = -λ[2:end].^2
+        K = @. ξ^2 + λ^2
 
-		K = I(N) * ξ^2 + diagm(0 => diagonal_elements,
-				       1 => above_diagonal_elements,
-				      -1 => below_diagonal_elements)
+        A = @. K / (K + μ) / ξ
 
-		A = (K / (K .+ diagm(μ))) / ξ
+    else
 
-	end	
+        # Calculate A in N-layer case (N > 1)
 
-	return A
+        diagonal_elements = [λ[1]^2; 2 * λ[2:end-1] .^ 2; λ[end]^2]
+        above_diagonal_elements = -λ[1:end-1] .^ 2
+        below_diagonal_elements = -λ[2:end] .^ 2
+
+        K =
+            I(N) * ξ^2 + diagm(
+                0 => diagonal_elements,
+                1 => above_diagonal_elements,
+                -1 => below_diagonal_elements,
+            )
+
+        A = (K / (K .+ diagm(μ))) / ξ
+
+    end
+
+    return A
 
 end
 
@@ -73,34 +76,37 @@ Arguments:
  - `μ`: nondimensional (y) vorticity gradient in each layer, Number or Vector
 """
 function B_func(ξ::Number, λ::Union{Vector,Number}, μ::Union{Vector,Number})
-	
-	N = length(μ)
 
-	if N == 1
+    N = length(μ)
 
-		# Calculate B in 1-layer case
+    if N == 1
 
-		K = @. ξ^2 + λ^2
+        # Calculate B in 1-layer case
 
-		B = @. 1 / (K + μ) / ξ
+        K = @. ξ^2 + λ^2
 
-	else
+        B = @. 1 / (K + μ) / ξ
 
-		# Calculate B in N-layer case (N > 1)
-		
-		diagonal_elements       = [λ[1]^2; 2*λ[2:end-1].^2; λ[end]^2]
-		above_diagonal_elements = -λ[1:end-1].^2
-		below_diagonal_elements = -λ[2:end].^2
+    else
 
-		K = I(N) * ξ^2 + diagm(0 => diagonal_elements,
-				       1 => above_diagonal_elements,
-				      -1 => below_diagonal_elements)
+        # Calculate B in N-layer case (N > 1)
 
-		B = inv(K .+ diagm(μ)) / ξ
+        diagonal_elements = [λ[1]^2; 2 * λ[2:end-1] .^ 2; λ[end]^2]
+        above_diagonal_elements = -λ[1:end-1] .^ 2
+        below_diagonal_elements = -λ[2:end] .^ 2
 
-	end
+        K =
[…]
