### 2025-02-27 bd053f5
update README with draft contributing info
