### 2025-02-27 mncrowe/QGDipoles.jl#31 @mncrowe
See PR here below. I suggest moving discussion to the PR so will close this issue.

https://github.com/mncrowe/QGDipoles.jl/pull/33
