### 2025-02-27 mncrowe/QGDipoles.jl#18 @mncrowe
Great, thanks! Will do this now.
