### 2025-02-27 9b731a2
Update Examples.md to hide `nothing` in example blocks
