+++ docs/src/Examples.md
@@ -27,7 +27,7 @@ tol = 1e-8	# maximum error in solution evaluation
 Nx, Ny = 512, 512
 Lx, Ly = 10, 10
 
-nothing
+nothing # hide
 ```
 
 We've taken ``M = 8`` as this is generally a sufficient number of terms to get a relative error ``< 10^{-6}`` in the final result.
@@ -44,7 +44,7 @@ We can now build the linear system and solve for the coefficients as follows:
 A, B, c, d = BuildLinSys(M, λ, μ; tol)
 K, a = SolveInhomEVP(A, B, c, d; K₀ = 4, tol)
 
-nothing
+nothing # hide
 ```
 
 The intermediate parameters, ``\lambda`` and ``\mu``, describe the rescaled vortex radius and PV gradient.
@@ -57,7 +57,7 @@ grid = CreateGrid(Nx, Ny, Lx, Ly)
 ψ, q = Calc_ψq(a, U, ℓ, R, β, grid)
 u, v = Calc_uv(ψ, grid)
 
-nothing
+nothing # hide
 ```
 
 We can plot our solution using Plots.jl:
@@ -108,7 +108,7 @@ tol = 1e-8			# maximum error in solution evaluation
 Nx, Ny = 512, 512
 Lx, Ly = [0, 10], [0, 10]
 
-nothing
+nothing # hide
 ```
 
 We've assumed that only the middle layer is active.
@@ -125,7 +125,7 @@ We start by building the full linear system:
 
 A, B, c, d = BuildLinSys(M, λ, μ; tol)
 
-nothing
+nothing # hide
 ```
 
 Next we remove the passive layers:
@@ -133,7 +133,7 @@ Next we remove the passive layers:
 ```@example multilayer
 A, B, c, d = ApplyPassiveLayers(A, B, c, d, ActiveLayers)
 
-nothing
+nothing # hide
 ```
 
 We can now solve the reduced system and put the passive layers, which have ``(K, \textbf{a}) = (0, \textbf{0})``, back in to ensure the sizes of ``K`` and ``\textbf{a}`` match the number of layers:
@@ -142,7 +142,7 @@ We can now solve the reduced system and put the passive layers, which have ``(K,
 K, a = SolveInhomEVP(A, B, c, d; K₀ = 4, tol)
 K, a = IncludePassiveLayers(K, a, ActiveLayers)
 
-nothing
+nothing # hide
 ```
 
 Finally, we can calculate our solution:
@@ -153,7 +153,7 @@ Finally, we can calculate our solution:
 grid = CreateGrid(Nx, Ny, Lx, Ly)
 ψ, q = Calc_ψq(a, U, ℓ, R, β, grid, x₀)
 
-nothing
+nothing # hide
 ```
 
 ## Example 3: SQG
@@ -187,7 +187,7 @@ sqg = true	    	# functions use SQG functionality
 Nx, Ny = 512, 512
 Lx, Ly = 10, 10
 
-nothing
+nothing # hide
 ```
 
 We have introduced a couple of new variables here.
@@ -206,7 +206,7 @@ Next we can build the linear system:
 A, B, c, d = BuildLinSys(M, λ, μ; tol, sqg)
 K, a = SolveInhomEVP(A, B, c, d; K₀ = 4, tol, method, sqg)
 
-nothing
+nothing # hide
 ```
 
 And finally we can create our solution:
@@ -218,7 +218,7 @@ grid = CreateGrid(Nx, Ny, Lx, Ly; cuda)
 ψ, b = Calc_ψb(a, U, ℓ, R, β, grid)
 u, v = Calc_uv(ψ, grid)
 
-nothing
+nothing # hide
[…]
