### 2025-02-27 mncrowe/QGDipoles.jl#23 @mncrowe
This was the original draft of the docs before moving to `Documenter.jl`. I'll remove it.
