+++ docs/readme.md
@@ -1,593 +1,0 @@
-# QGDipoles.jl
-
-Documentation and examples for QGDipoles.jl by Matthew N. Crowe.
-This document contains a summary of the package and some examples.
-See [here](https://mncrowe.github.io/QGDipoles.jl/) for the full documentation.
-
-## About
-
-This Julia package provides functions for evaluating dipolar vortex solutions in the surface quasi-geostrophic (SQG) and multi-layer quasi-geostrophic (LQG) models.
-It is intended for use by those researching vortex dynamics in strongly rotating flows, in particular for researchers in physical oceanography and atmospheric dynamics.
-This package is based on the semi-analytic theory of dipolar vortices derived in [1] and [2] for SQG solutions and [3] for LQG solutions.
-The method used a basis of orthogonal polynomials (Zernike radial functions) to convert a steady PDE into a linear algebra system which is solved using standard methods.
-This code consists of an updated version of the MATLAB code released as supplementary material with [3] and incorporates (unreleased) functions for the SQG problem.
-
-## Method Summary
-
-The full method is outlined in [1], [2] and [3].
-A summary is presented here such that the notation and examples presented later make some sense.
-We consider a dipolar vortex of radius $\ell$, moving with speed $U$.
-This vortex consists of an isolated region of high vorticity with a closed streamline at $x^2 + y^2 = \ell^2$ (in a frame co-moving with the vortex), hence fluid does not escape during propagation.
-Within the vortex core, $x^2 + y^2 < \ell^2$, are two counter rotating regions, corresponding to a dipole.
-The streamfunction describing the flow is denoted by $\psi$ and potential vorticity (PV) anomaly by $q$.
-Velocities may be derived as $(u, v) = (-\partial_y\psi, \partial_x\psi)$. The streamfunction, $\psi$, and PV anomaly, $q$, are related through PV inversion.
-In the case of multiple layers, $\psi$ and $q$ are vector valued functions of length equal to the number of layers, $N$.
-
-### Layered Quasi-Geostrophic (LQG) Solutions
-
-In the LQG model, steady, propagating, dipolar vortices satisfy the relation
-
-$$ q_i + \beta_i y = F_i(\psi_i + Uy),$$
-
-where $\beta_i$ denotes the background PV gradient, $i \in [1,\dots,N]$ is the layer index and $F_i$ is an arbitrary (piecewise continuous) function.
-To proceed, we assume that $F_i(z) = (\beta_i/U) z$ for $x^2 + y^2 > \ell^2$ (outside the vortex) and $F_i(z) = -(K_i^2/\ell^2) z$ for $x^2 + y^2 < \ell^2$ (inside the vortex).
-Using a Hankel transform and expansion in term of Zernike radial functions, the problem may be reduced to the linear algebra system
-
-$$ \left[ \textbf{A} -  \sum _{n = 1}^N K_n^2\textbf{B}_n \right] \textbf{a} = \textbf{c}_0 + \sum _{n = 1}^N K_n^2 \textbf{c}_n, $$
-
-where $\textbf{A}$ and $\textbf{B}_n$ are matrices, $\textbf{a}$ is a vector containing the coefficients in the polynomial expansion, $\textbf{c}_j$ are vectors and the $K_n$ are defined in $F_i$ above and appear as unknown eigenvalues in the linear problem.
-In order to solve the system, $N$ additional conditions are required.
-These are $\textbf{d}_n \cdot a = 0$ for $n \in [1, \dots, N]$ where the $\textbf{d}_n$ are vectors.
[…]
