### 2025-02-27 d630666
Update README.md with additional contribution information and guidelines
