### 2025-02-27 mncrowe/QGDipoles.jl#17 @mncrowe
Thanks, very useful. I've implemented these here:

https://github.com/mncrowe/QGDipoles.jl/pull/33
