### 2025-02-27 mncrowe/QGDipoles.jl#21 @mncrowe
Oops. Good catch.

There's a line in `CreateGrid` which is

```julia
Lx = Lx[2]
```

and should be

```julia
Lx = Lx[2] - Lx[1]
```

and similarly for `Ly`. Will do a PR.
