### 2025-02-27 mncrowe/QGDipoles.jl#20 @mncrowe
So a simple test that tries to build a few structures?

What counts as a failure? An error during construction or a failed comparison with other functionality (e.g. test vector inputs vs scalar inputs)?
