### 2025-02-27 mncrowe/QGDipoles.jl#22 @mncrowe
This is some great functionality I didn't know about. But I'm wondering about how to go about this with the examples. It seems that getting users to define this recipe could lead to more confusion when the basic plotting is fairly straightforward.

It might be worth adding to the main package but then we need `Plots.jl` as a dependency which I'd rather avoid as it isn't really needed.

Thoughts?
