### 2025-02-27 mncrowe/QGDipoles.jl#19 @mncrowe
As another option; maybe instructions for how to clone the examples folder, set up an environment and install the relevant packages?

This could go as a new first section on the examples page of the docs.
