### 2025-02-27 mncrowe/QGDipoles.jl#25 @mncrowe
Very helpful, will update all docstrings while reviewing function arguments.
