### 2025-02-27 mncrowe/QGDipoles.jl#22 @mncrowe
> Did that clarify my suggestion?

Yes, thanks. Didn't understand initially that Plots wasn't needed. Will implement this.
