### 2025-02-27 a38e164
Fix typos in paper.md
