### 2025-02-27 abbc219
Update paper.bib for consistency
