+++ paper/paper.bib
@@ -13,9 +13,9 @@
 }
 
 @software{FourierFlows,
-  author       = {Constantinou, N. C. and
-                  Wagner, L. C. and
-                  Palóczy, A.},
+  author       = {Navid C. Constantinou and
+  	             Gregory LeClaire Wagner and
+          	     Andr{{\'e}} Pal{{\'o}}czy}
   title        = {FourierFlows/FourierFlows.jl: v0.6.17},
   year         = {2021},
   publisher    = {Zenodo},
