+++ paper/paper.bib
@@ -15,7 +15,7 @@
 @software{FourierFlows,
   author       = {Navid C. Constantinou and
   	             Gregory LeClaire Wagner and
-          	     Andr{{\'e}} Pal{{\'o}}czy}
+          	     Andr{{\'e}} Pal{{\'o}}czy},
   title        = {FourierFlows/FourierFlows.jl: v0.6.17},
   year         = {2021},
   publisher    = {Zenodo},
