### 2025-02-27 mncrowe/QGDipoles.jl#29 @mncrowe
I've created a script to generate the figures. I've also replaced the existing figure to increase the text size.
