### 2025-02-27 07727f5
Add script to generate figures
