### 2025-02-27 0e2eae9
Add links to issues and discussions to assist new users
