### 2025-02-27 mncrowe/QGDipoles.jl#28 @mncrowe
I've started to make changes here:

https://github.com/mncrowe/QGDipoles.jl/pull/33

I suggest moving the discussion there though I'll keep this open until I've addressed all the points.
