### 2025-02-27 mncrowe/QGDipoles.jl#19 @haakon-e
That's a great idea!
