### 2025-02-27 mncrowe/QGDipoles.jl#20 @haakon-e
If you wrap the test in

```julia
@testset begin
  # [...]
end
```
then any failure, even those not explicitly `@test`ed will result in an error.

A very quick thing to do could be to include the files from the `examples/` folder in tests, just to verify that they run:
```Julia
using QGDipoles
ex_dir = joinpath(pkgdir(QGDipoles), "examples")
@testset "Test that examples run" begin
  include(joinpath(ex_dir, "1_Layer_Modon.jl"))
  # [...]
end
```

But yeah, ideally, there should be tests that verify basic (but non-trivial) code functionality. For the sake of this joss paper, you can keep it simple. Mostly, I think it'll save you and others for headaches in the future.
