### 2025-02-27 mncrowe/QGDipoles.jl#26 @haakon-e
> So this would also mean removing internal functions from the `export` list right?

Yes.

> And then importing them explicitly when required internally?

No. Since the entire package is one module, `QGDipoles`, all methods are already available for internal use. You only need to define (or `include`) internal functions before using them, I think.
