### 2025-02-28 3021715
fix bug with JuliaFormatter suggestion
Co-authored-by: Haakon Ludvig Langeland Ervik <45243236+haakon-e@users.noreply.github.com>
