### 2025-02-28 b2290f1
clarify documentation build
Co-authored-by: Haakon Ludvig Langeland Ervik <45243236+haakon-e@users.noreply.github.com>
