### 2025-02-28 d49e05b
Add additional info on documenter and formatter
