### 2025-02-28 703ba5f
require x,y as `AbstractVector` in `GridStruct`
