### 2025-02-28 e8a88f2
Improve installation instructions
