### 2025-02-28 957a1a8
add `import Pkg` to documentation guide
