+++ docs/reset_project.sh
@@ -1,0 +1,3 @@
+# Removes "QGDipoles.jl" from Project.toml
+
+sed -i '/QGDipoles/d' Project.toml
