+++ docs/src/Methodology.md
@@ -1,0 +1,204 @@
+# Methodology
+
+This page contains a summary of the numerical method used to determine steady dipolar solutions to quasi-geostrophic systems. A summary of the notation used is also included below.
+
+## Method Summary
+
+This package is based on the semi-analytic theory of dipolar vortices derived in Johnson & Crowe 2023[^1] and Crowe & Johnson 2023[^2] for SQG solutions and Crowe & Johnson 2024[^3] for LQG solutions.
+A summary is presented here such that the notation and examples presented elsewhere make some sense.
+We consider a dipolar vortex of radius ``\ell``, moving with speed ``U``.
+This vortex consists of an isolated region of high vorticity with a closed streamline at ``x^2 + y^2 = \ell^2`` (in a frame co-moving with the vortex), hence fluid does not escape during propagation. Within the vortex core, ``x^2 + y^2 < \ell^2``, are two counter rotating regions, corresponding to a dipole.
+The streamfunction describing the flow is denoted by ``\psi`` and potential vorticity (PV) anomaly by ``q``.
+Velocities may be derived as ``(u, v) = (-\partial_y\psi, \partial_x\psi)``.
+The streamfunction, ``\psi``, and PV anomaly, ``q``, are related through PV inversion. In the case of multiple layers, ``\psi`` and ``q`` are vector valued functions of length equal to the number of layers, ``N``.
+
+### Layered Quasi-Geostrophic (LQG) Solutions
+
+In the LQG model, steady, propagating, dipolar vortices satisfy the relation
+
+```math
+q_i + \beta_i y = F_i(\psi_i + Uy),
+```
+
+where ``\beta_i`` denotes the background PV gradient, ``i \in [1,\dots,N]`` is the layer index and ``F_i`` is an arbitrary (piecewise continuous) function.
+To proceed, we assume that ``F_i(z) = (\beta_i/U) z`` for ``x^2 + y^2 > \ell^2`` (outside the vortex) and ``F_i(z) = -(K_i^2/\ell^2) z`` for ``x^2 + y^2 < \ell^2`` (inside the vortex).
+Using a Hankel transform and expansion in term of Zernike radial functions, the problem may be reduced to the linear algebra system
+
+```math
+\left[ \textbf{A} -  \sum _{n = 1}^N K_n^2\textbf{B}_n \right] \textbf{a} = \textbf{c}_0 + \sum _{n = 1}^N K_n^2 \textbf{c}_n,
+```
+
+where ``\textbf{A}`` and ``\textbf{B}_n`` are matrices, ``\textbf{a}`` is a vector containing the coefficients in the polynomial expansion, ``\textbf{c}_j`` are vectors and the ``K_n`` are defined in ``F_i`` above and appear as unknown eigenvalues in the linear problem.
+In order to solve the system, ``N`` additional conditions are required.
+These are ``\textbf{d}_n \cdot \textbf{a} = 0`` for ``n \in [1, \dots, N]`` where the ``\textbf{d}_n`` are vectors.
+These conditions correspond to the requirement that the streamfunction and vorticity are continuous in each layer.
+In principal, we have an infinite number of coefficients in ``\textbf{a}``.
+However, since we know that these coefficients must decay with increasing index (since ``\psi`` is continuous), we can truncate the expansion after ``M`` terms.
+The resulting linear system is of size ``MN \times MN``.
+
[…]
