### 2025-02-28 448e1be
update installation instructions and finish index.md page
