+++ docs/src/index.md
@@ -4,19 +4,18 @@ Documentation and examples for QGDipoles.jl by [Matthew N. Crowe](https://mncrow
 
 ## What is QGDipoles.jl?
 
-`QGDipoles.jl` is a Julia package which provides functions for evaluating dipolar vortex solutions in the surface quasi-geostrophic (SQG) and multi-layer quasi-geostrophic (LQG) models.
-It is intended for use by those researching vortex dynamics in strongly rotating flows, in particular for researchers in physical oceanography and atmospheric dynamics.
+QGDipoles.jl is a Julia package which provides a series of functions that allow users to easily generate dipolar vortex solutions to two of the most commonly used QG models; the multi-layer QG (LQG) system and the surface QG (SQG) system.
+Solutions for the three-dimensional quasi-geostrophic model and magneto-quasi-geostrophic model will be implemented in the future.
 
-## What can I use QGDipoles.jl for?
+QGDipoles.jl is intended for use by those researching vortex dynamics in strongly rotating flows, in particular for researchers in physical oceanography and atmospheric dynamics.
+It may be used for semi-analytical studies of steady vortex solutions or to initialise a simulation with a steadily propagating, balanced vortex.
+These vortex solutions may be used as initial conditions in full primitive equation models as well as QG models since they exist in QG balance so will not generate strong initial transients.
 
-The primary purpose of this package is to create dipolar vortex solutions for 
-
-
-### Integration with GeophysicalFlows.jl
-
-This package is designed to work with the `TwoDGrid` structure from [FourierFlows.jl](https://fourierflows.github.io/FourierFlowsDocumentation/) and [GeophysicalFlows.jl](https://fourierflows.github.io/GeophysicalFlowsDocumentation/stable/)[^6].
-As such, these functions may be used to define initial conditions for layered and surface quasi-geostrophic simulations which may run on either CPUs or GPUs.
+QGDipoles.jl is designed to be consistent with the framework of [GeophysicalFlows.jl](https://fourierflows.github.io/GeophysicalFlowsDocumentation/stable/)[^4], a Julia package that contains modules for running LQG and SQG simulations on CPUs or GPUs.
+As such, QGDipoles.jl accepts grid inputs generated using the `TwoDGrid` function from [FourierFlows.jl](https://fourierflows.github.io/FourierFlowsDocumentation/) and can generate solution arrays on both CPUs and GPUs using CUDA.jl.
 However, [FourierFlows.jl](https://fourierflows.github.io/FourierFlowsDocumentation/) and [GeophysicalFlows.jl](https://fourierflows.github.io/GeophysicalFlowsDocumentation/stable/) are NOT required to use this package as an alternative grid structure is available (created using `CreateGrid`), which uses the same field names as the eqivalent [FourierFlows.jl](https://fourierflows.github.io/FourierFlowsDocumentation/) function.
+
+Full documentation exists with examples covering a range of LQG and SQG solutions and demonstrating how QGDipoles.jl may be integrated with [GeophysicalFlows.jl](https://fourierflows.github.io/GeophysicalFlowsDocumentation/stable/).
 
 ## How does QGDipoles.jl work?
 
@@ -28,3 +27,4 @@ For those interested in the original (LQG only) implementation, it can be found 
 [^1]: [Johnson, E. R., and M. N. Crowe, 2023, Oceanic dipoles in a surface quasigeostrophic model, J. Fluid Mech., 958, R2](https://doi.org/10.1017/jfm.2023.87).
[…]
