### 2025-02-28 mncrowe/QGDipoles.jl#25 @mncrowe
Addressed here: https://github.com/mncrowe/QGDipoles.jl/pull/35
