### 2025-02-28 94c10cb
Update docstrings and function list
