+++ src/create_modon.jl
@@ -54,15 +54,15 @@ GeophysicalFlows.jl so will also work with this grid.
 
 
 """
-Function: `ZernikeR(n, x)`
+    ZernikeR(n, x)
 
 Define the Zernike radial function using the `jacobi` function from SpecialFunctions
 
-Arguments:
+# Arguments:
  - `n`: order, Integer
  - `x`: evaluation point, Number or Array
 
-Note: this function is defined on [-1, 1] and is set to 0 for |x| > 1
+Note: this function is defined on ``[-1, 1]`` and is set to ``0`` for ``|x| > 1``
 """
 function ZernikeR(n::Int, x::Union{Number,Array})
 
@@ -72,11 +72,11 @@ function ZernikeR(n::Int, x::Union{Number,Array})
 end
 
 """
-Structure: `GridStruct`
+    GridStruct
 
 Stores the grid variables in physical and Fourier space
 
-Arguments:
+# Arguments:
  - `x`, `y`: x and y points in physical space, Ranges
  - `kr`, `l`: x and y points in Fourier space, Arrays
  - `Krsq`: `kr²+l²` in Fourier space, Array
@@ -93,11 +93,11 @@ struct GridStruct
 end
 
 """
-Function: `CreateGrid(Nx, Ny, Lx, Ly; cuda=false)`
+    CreateGrid(Nx, Ny, Lx, Ly; cuda=false)
 
 Define the numerical grid as a `GridStruct`
 
-Arguments:
+# Arguments:
  - `Nx`, `Ny`: number of gridpoints in x and y directions, Integers
  - `Lx`, `Ly`: x and y domains, either vectors of endpoints or lengths, Vectors or Numbers
  - `cuda`: `true`; use CUDA CuArray for fields (default: `false`)
@@ -157,11 +157,11 @@ function CreateGrid(
 end
 
 """
-Function: `Calc_ψq(a, U, ℓ, R, β, grid, x₀=[0, 0], α=0)`
+    Calc_ψq(a, U, ℓ, R, β, grid, x₀=[0, 0], α=0)
 
-Calculate ψ and q in a layered QG model using coefficients and vortex parameters
+Calculate ``ψ`` and ``q`` in a layered QG model using coefficients and vortex parameters
 
-Arguments:
+# Arguments:
  - `a`: M x N array of coefficients, Array
  - (`U`, `ℓ`): vortex speed and radius, Numbers
  - (`R`, `β`): Rossby radii and (y) PV gradients in each layer, Numbers or Vectors
@@ -265,16 +265,16 @@ function Calc_ψq(
 end
 
 """
-Function: `Calc_ψb(a, U, ℓ, R, β, grid, x₀=[0, 0], α=0)`
+    Calc_ψb(a, U, ℓ, R, β, grid, x₀=[0, 0], α=0)
 
-Calculate SQG fields ψ and b using coefficients and vortex parameters
+Calculate SQG fields ``ψ`` and ``b`` using coefficients and vortex parameters
 
-Arguments:
+# Arguments:
  - `a`: M x 1 array of coefficients, Array
  - (`U`, `ℓ`): vortex speed and radius, Numbers
- - `R`: vector of [R, R'], Vector
+ - `R`: vector of ``[R, R']``, Vector
  - `β`: beta-plane (y) PV gradient, Number
- - `grid`: grid structure containing x, y, and Krsq
+ - `grid`: grid structure containing `x`, `y`, and `Krsq`
  - `x₀`: position of vortex center, vector (default: `[0, 0]`)
  - `α`: initial angle of vortex, Number (default: `0`)
 
[…]
