+++ src/vortex_types.jl
@@ -7,11 +7,11 @@ solution to be included in the same structure and allow parameters to be checked
 """
 
 """
-Structure: `LQGParams`
+    LQGParams
 
 Stores the parameters for an LQG dipolar vortex solution
 
-Arguments:
+# Arguments:
  - `U`: Vortex speed
  - `ℓ`: Vortex radius
  - `R`: Rossby radius
@@ -51,11 +51,11 @@ struct LQGParams
 end
 
 """
-Structure: `SQGParams`
+    SQGParams
 
 Stores the parameters for an SQG dipolar vortex solution
 
-Arguments:
+# Arguments:
  - `U`: Vortex speed
  - `ℓ`: Vortex radius
  - `R`: Rossby radius
@@ -88,11 +88,11 @@ struct SQGParams
 end
 
 """
-Structure: `LQGVortex`
+    LQGVortex
 
 Stores fields and diagnostics for an LQG dipolar vortex solution
 
-Arguments:
+# Arguments:
  - `params`: Vortex params
  - `ψ`: streamfunction
  - `q`: potential vorticity anomaly
@@ -120,11 +120,11 @@ struct LQGVortex
 end
 
 """
-Structure: `SQGVortex`
+    SQGVortex
 
 Stores fields and diagnostics for an SQG dipolar vortex solution
 
-Arguments:
+# Arguments:
  - `params`: Vortex params
  - `ψ`: surface streamfunction
  - `b`: surface buoyancy
@@ -150,11 +150,11 @@ struct SQGVortex
 end
 
 """
-Function: `DefLQGParams`
+    DefLQGParams(U=1, ℓ=1, R=Inf, β=0, ActiveLayers=1, H=1, x₀=[0,0], α=0, M=8; tol=1e-6, K₀=nothing, a₀=nothing, UseAnalytic=false, CalcVelocity=false, CalcVorticity=false, CalcEnergy=false, CalcEnstrophy=false)
 
-Defines an LQGParams structure using the given inputs
+Defines an `LQGParams` structure using the given inputs
 
-Arguments:
+# Arguments:
  - `U`: Vortex speed
  - `ℓ`: Vortex radius
  - `R`: Rossby radius
@@ -215,11 +215,11 @@ function DefLQGParams(
 end
 
 """
-Function: `DefSQGParams`
+    DefSQGParams(U=1, ℓ=1, R=[Inf,Inf], β=0, x₀=[0,0], α=0, M=12; tol=1e-6, K₀=nothing, a₀=nothing, CalcVelocity=false, CalcVorticity=false, CalcEnergy=false)
 
-Defines an SQGParams structure using the given inputs
+Defines an `SQGParams` structure using the given inputs
 
-Arguments:
+# Arguments:
  - `U`: Vortex speed
  - `ℓ`: Vortex radius
  - `R`: Rossby radius
@@ -268,11 +268,12 @@ function DefSQGParams(
 end
 
 """
-Function: `DefLQGVortex`
+    DefLQGVortex(grid, U=1, ℓ=1, R=Inf, β=0, ActiveLayers=1, H=1, x₀=[0,0], α=0, M=8; tol=1e-6, K₀=nothing, a₀=nothing, UseAnalytic=false, CalcVelocity=false, CalcVorticity=false, CalcEnergy=false, CalcEnstrophy=false)
 
-Defines an LQGVortex solution structure using the given inputs
 
-Arguments:
+Defines an `LQGVortex` solution structure using the given inputs
+
+# Arguments:
  - `grid`: grid structure
  - `U`: Vortex speed
  - `ℓ`: Vortex radius
[…]

[…]
