### 2025-02-28 mncrowe/QGDipoles.jl#26 @mncrowe
I've moved internal functions/types to a section at the end and grouped the rest by functionality rather than source file. Changes in PR: https://github.com/mncrowe/QGDipoles.jl/pull/35
