### 2025-02-28 759008e
implement recipe for plotting and simplify plotting in example scripts
