+++ examples/3_Layer_Modon.jl
@@ -39,14 +39,4 @@ grid = CreateGrid(Nx, Ny, Lx, Ly; cuda)
 
 using Plots
 
-heatmap(
-    grid.x,
-    grid.y,
-    transpose(ψ[:, :, 2]);
-    colormap = :balance,
-    aspect_ratio = 1,
-    xlims = Lx,
-    ylims = Ly,
-    xlabel = "x",
-    ylabel = "y",
-)
+heatmap(grid, ψ, layer = 2)
