+++ docs/reset_project.sh
@@ -1,3 +1,0 @@
-# Removes "QGDipoles.jl" from Project.toml
-
-sed -i '/QGDipoles/d' Project.toml
