+++ test/Type_tests.jl
@@ -1,0 +1,5 @@
+"""
+Tests for the type/structures from QGDipoles.jl.
+"""
+
+
