+++ examples/SQG_Modon.jl
@@ -40,14 +40,4 @@ u, v = Calc_uv(ψ, grid)
 
 using Plots
 
-heatmap(
-    grid.x,
-    grid.y,
-    transpose(b);
-    colormap = :balance,
-    aspect_ratio = 1,
-    xlims = (-Lx / 2, Lx / 2),
-    ylims = (-Ly / 2, Ly / 2),
-    xlabel = "x",
-    ylabel = "y",
-)
+heatmap(grid, b)
