### 2025-02-28 mncrowe/QGDipoles.jl#22 @mncrowe
This suggestion worked great. I've included this in an PR: https://github.com/mncrowe/QGDipoles.jl/pull/35

I've updated all example scripts to use this recipe. Examples in the docs will be updated when I do a full re-write of the examples page (upcoming).
