### 2025-02-28 934a7ef
test all examples (excluding those related to GeophysicalFlows) and add recipes to docs
