### 2025-02-28 mncrowe/QGDipoles.jl#20 @mncrowe
I've included all examples in the tests, excluding any that rely on GeophysicalFlows as I don't want that as a dependency. As a consequence I've had to comment out the plotting lines with a comment saying these can be uncommented if you have `Plots` installed. I think this is fine as experienced users can just run the script and use their preferred plotting package and inexperienced ones can follow the (upcoming) guide on how to add Plots.

I'll add a few extra tests to test any untested conditional functionality and the vortex structures (which are newish and haven't got proper tests).
