### 2025-02-28 e703275
Add functions to update vortex types and new examples for vortex types
