+++ src/vortex_types.jl
@@ -7,6 +7,24 @@ solution to be included in the same structure and allow parameters to be checked
 """
 
 """
+    CreateGrid(; Nx=512, Ny=512, Lx=[-5,5], Ly=[-5,5], cuda=false)
+
+Define the numerical grid as a `GridStruct`
+
+# Arguments:
+ - `Nx`, `Ny`: number of gridpoints in x and y directions, Integers
+ - `Lx`, `Ly`: x and y domains, either vectors of endpoints or lengths, Vectors or Numbers
+ - `cuda`: `true`; use CUDA CuArray for fields (default: `false`)
+"""
+CreateGrid(;
+    Nx = 512,
+    Ny = 512,
+    Lx = [-5, 5],
+    Ly = [-5, 5],
+    cuda = false,
+) = CreateGrid(Nx, Ny, Lx, Ly; cuda)
+
+"""
     LQGParams
 
 Stores the parameters for an LQG dipolar vortex solution
@@ -150,7 +168,7 @@ struct SQGVortex
 end
 
 """
-    DefLQGParams(U=1, ℓ=1, R=Inf, β=0, ActiveLayers=1, H=1, x₀=[0,0], α=0, M=8; tol=1e-6, K₀=nothing, a₀=nothing, UseAnalytic=false, CalcVelocity=false, CalcVorticity=false, CalcEnergy=false, CalcEnstrophy=false)
+    DefLQGParams(; U=1, ℓ=1, R=Inf, β=0, ActiveLayers=1, H=1, x₀=[0,0], α=0, M=8, tol=1e-6, K₀=nothing, a₀=nothing, UseAnalytic=false, CalcVelocity=false, CalcVorticity=false, CalcEnergy=false, CalcEnstrophy=false)
 
 Defines an `LQGParams` structure using the given inputs
 
@@ -173,7 +191,7 @@ Defines an `LQGParams` structure using the given inputs
  - `CalcEnergy`: flag to determine if energy is calculated
  - `CalcEnstrophy`: flag to determine if enstrophy is calculated
 """
-function DefLQGParams(
+function DefLQGParams(;
     U::Number = 1,
     ℓ::Number = 1,
     R::Union{Number,Vector} = Inf,
@@ -182,7 +200,7 @@ function DefLQGParams(
     H::Union{Number,Vector} = 1,
     x₀::Vector = [0, 0],
     α::Number = 0,
-    M::Int = 8;
+    M::Int = 8,
     tol::Number = 1e-6,
     K₀::Union{Number,Array,Nothing} = nothing,
     a₀::Union{Array,Nothing} = nothing,
@@ -212,10 +230,11 @@ function DefLQGParams(
         CalcEnergy,
         CalcEnstrophy,
     )
+
 end
 
 """
-    DefSQGParams(U=1, ℓ=1, R=[Inf,Inf], β=0, x₀=[0,0], α=0, M=12; tol=1e-6, K₀=nothing, a₀=nothing, CalcVelocity=false, CalcVorticity=false, CalcEnergy=false)
+    DefSQGParams(; U=1, ℓ=1, R=[Inf,Inf], β=0, x₀=[0,0], α=0, M=12, tol=1e-6, K₀=nothing, a₀=nothing, CalcVelocity=false, CalcVorticity=false, CalcEnergy=false)
 
 Defines an `SQGParams` structure using the given inputs
 
@@ -234,14 +253,14 @@ Defines an `SQGParams` structure using the given inputs
  - `CalcVorticity`: flag to determine if vorticity is calculated
  - `CalcEnergy`: flag to determine if energy is calculated
 """
-function DefSQGParams(
+function DefSQGParams(;
     U::Number = 1,
[…]
