### 2025-02-28 mncrowe/QGDipoles.jl#24 @mncrowe
I'm in the process of updating the examples. The new recommended way to use the package is via the vortex structures which are (now) defined by keyword-based functions.

I intend to do a similar pass over what I've labelled as the 'high level' functions. The 'low level' and 'internal' functions will likely remain as they are.
