+++ src/LQG/lqg_evp.jl
@@ -58,7 +58,7 @@ the LQG problem
  - `μ`: nondimensional (y) vorticity gradient in each layer, Number or Vector
  - `tol`: error tolerance for QuadGK via `JJ_int`, Number (default: `1e-6`)
 """
-function BuildLinSys(
+function BuildLinSysLQG(
     M::Int,
     λ::Union{Vector,Number},
     μ::Union{Vector,Number};
