### 2025-03-01 7310091
convert user-facing functions to be keyword based and add tests, definitely broken
