+++ examples/Low_Level/N_Layer_Modon.jl
@@ -33,8 +33,8 @@ K, a = IncludePassiveLayers(K, a, ActiveLayers)
 # Create grid and calculate streamfunctions and vorticities
 
 grid = CreateGrid(Nx, Ny, Lx, Ly; cuda)
-ψ, q = Calc_ψq(a, U, ℓ, R, β, grid)
-u, v = Calc_uv(ψ, grid)
+ψ, q = Calc_ψq(grid, a; U, ℓ, R, β)
+u, v = Calc_uv(grid, ψ)
 
 # Plot streamfunction ψ in layer 1, if we have `Plots.jl` added
 
