### 2025-03-01 mncrowe/QGDipoles.jl#20 @mncrowe
I've completely rearranged the testing to be consistent with the rearranged codebase.

In particular:
* All structure construction is now tested
* Functions with conditions on inputs are tested for each input case
* Some previously untested functions are now tested

PR here: https://github.com/mncrowe/QGDipoles.jl/pull/35

Will close this and move further discussion of tests to PR
