### 2025-03-01 19cd2b7
restructure tests and add tests for untested features
