+++ test/LQG_tests.jl
@@ -1,5 +1,5 @@
 """
-Tests for the LQG functiond for QGDipoles.jl
+Tests for the LQG functions for QGDipoles.jl
 
 """
 
@@ -67,7 +67,7 @@ end
 Tests the value of maximum value of ψ for the LCD against a known value.
 
 """
-function Test1QG_ψ(cuda::Bool)
+function Test1QG_ψ(grid)
 
     # Define parameters
 
@@ -78,9 +78,6 @@ function Test1QG_ψ(cuda::Bool)
 
     λ, μ = 0, 0
 
-    Nx, Ny = 128, 128
-    Lx, Ly = 10, 10
-
     # Build linear system
 
     A, B, c, d = BuildLinSysLQG(M, λ, μ; tol)
@@ -88,7 +85,6 @@ function Test1QG_ψ(cuda::Bool)
 
     # Calculate solution
 
-    grid = CreateGrid(Nx, Ny, Lx, Ly; cuda = cuda)
     ψ, q = Calc_ψq(grid, a; U = 1, ℓ = 1, R = Inf, β = 0)
 
     return abs(maximum(ψ) - 1.28110) < 1e-5
@@ -131,14 +127,16 @@ function Test2QG_K()
 end
 
 """
-    Test2QG_PVinv(cuda)
+    Test2QG_PVinv(grid)
 
 Tests if the 3D PV inversion array has been built correctly to avoid Inf/NaN values in ψ.
 Common error in (0, 0) wavenumber due to divide by 0 issues. Due to differences in function
 support, these arrays are constructed differently on the CPU and GPU.
 
+Also test length(β) < N case through use of β = 0 for N = 2.
+
 """
-function Test2QG_PVinv(cuda)
+function Test2QG_PVinv(grid)
 
     # Define parameters
 
@@ -149,13 +147,10 @@ function Test2QG_PVinv(cuda)
 
     U, ℓ = 1, 1
     R = [Inf, 1]
-    β = [0, 1]
+    β = 0
 
     λ = ℓ ./ R
-    μ = β * ℓ^2 / U
-
-    Nx, Ny = 128, 128
-    Lx, Ly = 10, 10
+    μ = [β, β] * ℓ^2 / U
 
     # Build linear system
 
@@ -164,7 +159,6 @@ function Test2QG_PVinv(cuda)
 
     # Calculate solution
 
-    grid = CreateGrid(Nx, Ny, Lx, Ly; cuda)
     ψ, q = Calc_ψq(grid, a; U, ℓ, R, β)
 
     # Look for NaN and Inf values
@@ -227,3 +221,198 @@ function TestLQG_ActiveLayers()
     return If_K & If_N
 
 end
+
+"""
+    TestWrapperLQG(grid; U, ℓ, R, β)
+
+Tests the LQG wrapper by comparing against the results without a wrapper.
+
+"""
+function TestWrapperLQG(
+    grid;
+    U::Number,
+    ℓ::Number,
+    R::Union{Number,Vector},
+    β::Union{Number,Vector},
+)
+
+    # Set numerical parameters
+
+    M = 8
+    tol = 1e-8
+
+    # Calculate solution without wrapper
+
+    λ = ℓ ./ R
+    μ = β * ℓ^2 / U
+
+    A, B, c, d = BuildLinSysLQG(M, λ, μ; tol)
+    K, a = SolveInhomEVP(A, B, c, d; tol)
+    ψ₁, _ = Calc_ψq(grid, a; U, ℓ, R, β)
+
+    # Calculate solution with wrapper
+
[…]
