### 2025-03-01 mncrowe/QGDipoles.jl#24 @mncrowe
All user facing functions have been refactored as follows:
* `grid` is first argument (if required) and compulsory
* Compulsory field/array inputs are next (e.g. ψ for energy calculations)
* All other arguments are Keyword based and appear in the same order in all functions

This has been implemented here: https://github.com/mncrowe/QGDipoles.jl/pull/35

I'm going to close this and suggest moving any further discussion to that PR
