+++ src/extras/monopoles.jl
@@ -24,12 +24,7 @@ Note: This function outputs ``(u, v)`` directly since the solution has discontin
 the vortex boundary, ``r = ℓ``, so derivatives evaluated with Fourier transforms exhibit Gibbs
 phenomenon.
 """
-function CreateRankine(
-    grid;
-    ℓ::Number = 1,
-    Γ::Number = 2π,
-    x₀::Vector = [0, 0],
-)
+function CreateRankine(grid; ℓ::Number = 1, Γ::Number = 2π, x₀::Vector = [0, 0])
 
     # Create Cartesian and polar grids
 
@@ -167,11 +162,7 @@ Note: This function is designed to be used for creating periodic streamfunctions
 the vorticity fields generated by `CreateMonopole`. It does not support multi-layer QG
 and is only valid on an f-plane (β = 0).
 """
-function InvertVorticity1LQG(
-    grid,
-    q::Union{CuArray, Array};
-    R::Number = Inf,
-)
+function InvertVorticity1LQG(grid, q::Union{CuArray,Array}; R::Number = Inf)
 
     Nx = size(q, 1)
 
@@ -216,8 +207,8 @@ split into a dipole equation ``qⱼ + βⱼ = -Kⱼ²(ψⱼ + Uy)`` and a monopo
 function CreateLQGMonopole(
     grid;
     ℓ::Number = 1,
-    E::Union{Vector, Number} = 1,
-    R::Union{Vector, Number} = Inf,
+    E::Union{Vector,Number} = 1,
+    R::Union{Vector,Number} = Inf,
     x₀::Vector = [0, 0],
 )
 
@@ -255,11 +246,7 @@ This function inverts the potential vorticity relation ``q = ΔN ψ`` for the LQ
  - `R`: Rossby radius, Number or Vector (default: `Inf`)
 
 """
-function InvertVorticityLQG(
-    grid,
-    q::Union{CuArray, Array};
-    R::Union{Vector, Number} = Inf,
-)
+function InvertVorticityLQG(grid, q::Union{CuArray,Array}; R::Union{Vector,Number} = Inf)
 
     Nx, Ny = size(q)
     N = Int(length(q) / (Nx * Ny))
