+++ docs/src/Examples.md
@@ -3,15 +3,173 @@
 Here we present some examples which demonstrate how to use this package.
 Further examples are available in the `examples/` directory [here](https://github.com/mncrowe/QGDipoles.jl/tree/main/examples).
 
-## Example 1: 1-layer QG
+## Getting Started
+
+The first step is to create a new project (also known as an environment) and install Julia.
+We can do this by opening Julia with the current directory enabled using
+```shell
+julia --project=.
+```
+Further information on working with environments can be found [here](https://pkgdocs.julialang.org/v1/environments/).
+
+We can now install `QGDipoles.jl` to this project using the instructions on the Installation page.
+It is strongly recommended to install `QGDipoles.jl` and all other packages required for a particular project to a new environment for that project.
+
+To run the examples on this page we'll also need `Plots.jl` which can be installed using
+```julia
+import Pkg
+Pkg.add("Plots")
+Pkg.instantiate()
+```
+One of the examples below also required `GeophysicalFlows.jl` which can be installed in the same way.
+
+We can now follow the examples below or copy the `examples/` directory to our current working directory using
+```julia
+using QGDipoles
+cp(joinpath(pkgdir(QGDipoles), "examples"), "examples")
+```
+
+## Vortex Structures
+
+The recommended way to use `QGDipoles.jl` for new users is to create vortex structures.
+Let's create a single-layer QG vortex with unit radius and velocity ``ℓ = U = 1`` and impose a background rotation gradient of ``β = 1``.
+We'll start by creating a grid by specifying the number of gridpoints `(Nx, Ny)` and the domain size `(Lx, Ly)`
+```@example structures
+using QGDipoles, Plots
+
+grid = CreateGrid(; Nx = 512, Ny = 512, Lx = 10, Ly = 10)
+```
+We can now use this grid to create our vortex
+```@example structures
+vortex = DefLQGVortex(grid; U = 1, ℓ = 1, β = 1)
+```
+This function call may take a few seconds as Julia functions are [compiled before they're used for the first time](https://en.wikipedia.org/wiki/Just-in-time_compilation).
+Subsequent calls of the same function will be much faster.
+Vortex parameters are passed as [keyword arguments](https://docs.julialang.org/en/v1/manual/functions/#Keyword-Arguments), the full list of keyword arguments available for each function can be found on the [List of Functions](@ref Functions) page.
+Strictly, we don't need to include `ℓ` and `U` here as `1` is the default for these keywords.
+Our vortex solution can be plotted as a `heatmap` using
+```@example structures
+heatmap(grid, vortex.ψ)
+```
+We can now modify our vortex by including a finite Rossby radius of deformation (corresponding to the equivalent barotropic model) given by `R = 1`
+```@example structures
+vortex = UpdateVortex(grid, vortex; R = 1)
+```
+All other parameters remain unchanged and can be checked using
+```@example structures
+vortex.params
+```
+
+We can construct two-layer QG vortices and SQG vortices using the same approach
+```@example structures
+vortex_LQG = DefLQGVortex(grid; R = [1, 1], β = [1, 0])
+
+vortex_SQG = DefSQGVortex(grid)
+
+nothing # hide
+```
+We can plot layers from multi-layer vortices using the `layer` keyword in `heatmap`.
+
[…]
