+++ docs/src/Methodology.md
@@ -15,19 +15,15 @@ The streamfunction, ``\psi``, and PV anomaly, ``q``, are related through PV inve
 ### Layered Quasi-Geostrophic (LQG) Solutions
 
 In the LQG model, steady, propagating, dipolar vortices satisfy the relation
-
 ```math
 q_i + \beta_i y = F_i(\psi_i + Uy),
 ```
-
 where ``\beta_i`` denotes the background PV gradient, ``i \in [1,\dots,N]`` is the layer index and ``F_i`` is an arbitrary (piecewise continuous) function.
 To proceed, we assume that ``F_i(z) = (\beta_i/U) z`` for ``x^2 + y^2 > \ell^2`` (outside the vortex) and ``F_i(z) = -(K_i^2/\ell^2) z`` for ``x^2 + y^2 < \ell^2`` (inside the vortex).
 Using a Hankel transform and expansion in term of Zernike radial functions, the problem may be reduced to the linear algebra system
-
 ```math
 \left[ \textbf{A} -  \sum _{n = 1}^N K_n^2\textbf{B}_n \right] \textbf{a} = \textbf{c}_0 + \sum _{n = 1}^N K_n^2 \textbf{c}_n,
 ```
-
 where ``\textbf{A}`` and ``\textbf{B}_n`` are matrices, ``\textbf{a}`` is a vector containing the coefficients in the polynomial expansion, ``\textbf{c}_j`` are vectors and the ``K_n`` are defined in ``F_i`` above and appear as unknown eigenvalues in the linear problem.
 In order to solve the system, ``N`` additional conditions are required.
 These are ``\textbf{d}_n \cdot \textbf{a} = 0`` for ``n \in [1, \dots, N]`` where the ``\textbf{d}_n`` are vectors.
@@ -42,26 +38,20 @@ In the one-layer case the problem reduces to known analytical solutions, such as
 ### Surface Quasi-Geostrophic (SQG) Solutions
 
 In the SQG model, steady, propagating, dipolar vortices satisfy the relation
-
 ```math
 \left[\partial_z + \frac{1}{R'}\right] \psi = F(\psi + Uy),
 ```
-
 where
-
 ```math
 \partial_z = \sqrt{-\nabla^2 + \beta/U} \hspace{5pt} \tanh \left[R \sqrt{-\nabla^2 + \beta/U} \right],
 ```
-
 is a Dirichlet-Neumann operator linking the surface streamfunction, ``\psi``, and the surface buoyancy, ``b = \partial_z \psi``.
 Here, ``(R, R')`` describes the baroclinic and barotropic Rossby radii and ``\beta`` is the background vorticity gradient.
 We assume that ``F(z) = 0`` for ``x^2 + y^2 > \ell^2`` (outside the vortex) and ``F_i(z) = -(K/\ell) z`` for ``x^2 + y^2 < \ell^2``.
 Using a Hankel transform and expansion in term of Zernike radial functions, the problem may be reduced to the linear algebra system
-
 ```math
 \left[ \textbf{A} -  K\textbf{B} \right] \textbf{a} = \textbf{c}_0 + K \textbf{c}_1,
 ```
-
 where ``\textbf{A}`` and ``\textbf{B}`` are matrices, ``\textbf{c}_i`` are vectors, ``\textbf{a}`` is a vector of coefficients and ``K`` is an eigenvalue related to ``F``.
 An additional condition is required to solve this system for a unique set of ``K``.
 This condition is taken to be continuity across the vortex boundary and corresponds to ``\textbf{d} \cdot \textbf{a} = 0`` for some vector ``\textbf{d}``.
[…]
