### 2025-03-02 mncrowe/QGDipoles.jl#28 @mncrowe
Draft changes made for all points in PR. Will close this for now and move discussion to PR.
