### 2025-03-02 mncrowe/QGDipoles.jl#19 @mncrowe
> That's a great idea!

This idea has been implemented in latest PR: https://github.com/mncrowe/QGDipoles.jl/pull/35

Will close this and move discussion to PR.
