### 2025-03-19 08b39cd
fix colorbar in examples.md and add discussion of utils to Readme
