### 2025-03-19 db203e9
change arguments to fields in docstrings for structs and seperate argument and keyword arguments in all functions
