### 2025-03-19 9cdd475
add references to other functions in docstrings
