+++ src/SQG/sqg_evp.jl
@@ -44,7 +44,7 @@ the SQG problem
  - `μ`: nondimensional (y) vorticity gradient in each layer, Number or Vector
 
 # Keyword arguments:
- - `tol`: error tolerance for QuadGK via `JJ_int`, Number (default: `1e-6`)
+ - `tol`: error tolerance for `QuadGK` via `JJ_int`, Number (default: `1e-6`)
 """
 function BuildLinSysSQG(
     M::Int,
