+++ src/extras/monopoles.jl
@@ -71,9 +71,9 @@ for ``r > ℓ`` and ``z₀ = ψ(r=ℓ)``. These solutions exist only on an f-pla
  - `R`: Rossby radius (default: `Inf`)
  - `x₀`: position of vortex center, vector (default: `[0, 0]`)
 
-Note: This vortex has a continuous vorticity distribution so calculating (u, v)
-from ψ with Fourier transforms will work. This function outputs (u, v) from the
-analytical expressions for consistency with `CreateRankine`.
+Note: This vortex has a continuous vorticity distribution so calculating `(u, v)`
+from `ψ` with Fourier transforms will work. This function outputs `(u, v)` from the
+analytical expressions for consistency with [`CreateRankine`](@ref).
 """
 function Create1LMonopole(
     grid;
@@ -165,7 +165,7 @@ This function inverts the potential vorticity relation ``q = [∇²-1/R²]ψ`` f
  - `R`: Rossby radius (default: `Inf`)
 
 Note: This function is designed to be used for creating periodic streamfunctions using
-the vorticity fields generated by `CreateMonopole`. It does not support multi-layer QG
+the vorticity fields generated by [`CreateMonopole`](@ref). It does not support multi-layer QG
 and is only valid on an f-plane (β = 0).
 """
 function InvertVorticity1LQG(grid, q::Union{CuArray,Array}; R::Number = Inf)
