+++ src/extras/monopoles.jl
@@ -165,7 +165,7 @@ This function inverts the potential vorticity relation ``q = [∇²-1/R²]ψ`` f
  - `R`: Rossby radius (default: `Inf`)
 
 Note: This function is designed to be used for creating periodic streamfunctions using
-the vorticity fields generated by [`CreateMonopole`](@ref). It does not support multi-layer QG
+the vorticity fields generated by [`Create1LMonopole`](@ref). It does not support multi-layer QG
 and is only valid on an f-plane (β = 0).
 """
 function InvertVorticity1LQG(grid, q::Union{CuArray,Array}; R::Number = Inf)
