### 2025-03-19 ed36acd
add gpu as tip
Co-authored-by: Haakon Ludvig Langeland Ervik <45243236+haakon-e@users.noreply.github.com>
