### 2025-03-19 9871e30
update SQG example, replace integer  with symbol
