### 2025-03-19 470a0f9
fix examples based on changes
