### 2025-03-19 7b0db0d
run formatter and remove unneeded dependencies
