### 2025-03-24 #7767 @rwegener2
Thanks for the update @smillerc. Do you think that "Quality of Writing" is sufficient for acceptance? If so, would you mind checking that box on your checklist to confirm that?
