### 2025-03-25 f550e7f
Add initial files for 3D QG functions, implement 3D grid
