+++ src/3DQG/3dqg_grid.jl
@@ -1,0 +1,201 @@
+"""
+This file contains functions for building the 3D grid using a Fourier basis in the horizontal directions
+and Chebyshev points (with spectral collocation) in the vertical direction.
+
+...
+
+"""
+
+"""
+    GridChebyshev(N, L)
+
+Creates a grid of Chebyshev points of the second kind and a spectral collocation
+differentiation matrix.
+
+# Arguments:
+ - `N`: number of gridpoints
+ - `L`: vector of endpoints, ``z ∈ [L₁, L₂]``
+
+"""
+
+function GridChebyshev(N::Int, L::Vector)
+
+    z = (L[1] + L[2]) / 2 .- (L[2] - L[1]) / 2 * cos.((0:(N-1))*π/(N-1))
+
+    D = z' .- z
+
+    D₁ = D
+    D₁[D .== 0] .= 1
+
+    A = ones(N, 1) * prod(D₁; dims = 1)
+
+    M = A' ./ (A .* D')
+    M[1:(N+1):N^2] .= sum(1 ./ D₁; dims = 1)' .- 1
+
+    return z, M
+
+end
+
+"""
+    GridStruct3D
+
+Stores the 3D grid variables in physical and Fourier space
+
+# Fields:
+ - `x`, `y`: x and y points in physical space, Ranges
+ - `kr`, `l`: x and y points in Fourier space, Arrays
+ - `Krsq`: `kr²+l²` in Fourier space, Array
+ - `z`: z points in physical space, Vector
+ - `M`: spectral collocation matrix for differentiation in z, Matrix
+"""
+struct GridStruct3D
+    # position ranges for x, y and z
+    x::AbstractVector
+    y::AbstractVector
+    z::Vector
+    # wavenumber arrays in Fourier space
+    kr::Union{Array{Float64},CuArray{Float64}}
+    l::Union{Array{Float64},CuArray{Float64}}
+    # K² = kr²+l² array in Fourier space
+    Krsq::Union{Array{Float64},CuArray{Float64}}
+    # spectral collocation matrix in z
+    M::Union{Array{Float64},CuArray{Float64}}
+end
+
+"""
+    CreateGrid3D(Nx, Ny, Nz, Lx, Ly, Lz; cuda=false)
+
+Define the numerical grid as a [`GridStruct3D`](@ref)
+
+# Arguments:
+ - `Nx`, `Ny`, `Nz`: number of gridpoints in x and y directions, Integers
+ - `Lx`, `Ly`, `Lz`: x and y domains, either vectors of endpoints or lengths, Vectors or Numbers
+
+# Keyword arguments:
+ - `cuda`: `true`; use CUDA CuArray for fields (default: `false`)
+"""
+function CreateGrid3D(
+    Nx::Int,
+    Ny::Int,
+    Nz::Int,
+    Lx::Union{Number,Vector},
+    Ly::Union{Number,Vector},
+    Lz::Union{Number,Vector};
+    cuda::Bool = false,
+)
+
+    if length(Lx) == 2
+
+        x₀ = Lx[1]
+        Lx = Lx[2] - Lx[1]
+
+    else
+
+        x₀ = -Lx / 2
+
+    end
+
+    if length(Ly) == 2
+
+        y₀ = Ly[1]
+        Ly = Ly[2] - Ly[1]
+
+    else
+
+        y₀ = -Ly / 2
+
+    end
+
+    if length(Lz) == 2
+
+        z₀ = Lz[1]
+        Lz = Lz[2] - Lz[1]
+
+    else
+
+        z₀ = -Lz
+
+    end
+
+    Δx = Lx / Nx
+    Δy = Ly / Ny
+
[…]
