+++ src/3DQG/3dqg_evp.jl
@@ -1,0 +1,7 @@
+"""
+This file contains functions for building the eigenvalue problem for the 3D QG model
+
+...
+
+"""
+
