### 2025-04-01 #7767 @haakon-e
@rwegener2, @mncrowe Thanks for your patience.

I've now looked through @mncrowe's latest changes in the response to my suggestions in https://github.com/mncrowe/QGDipoles.jl/pull/35.

Thank you for thoroughly engaging with my comments and suggestions. I fully support publication of the paper. This has been a great review process and I wish you all the best for your software!
