+++ paper/paper.bib
@@ -175,11 +175,12 @@
 }
 
 @misc{CroweSutyrin,
-      title={Symmetry breaking of two-layer eastward propagating dipoles}, 
-      author={Matthew N. Crowe and Georgi G. Sutyrin},
-      year={2024},
-      eprint={2410.14402},
-      archivePrefix={arXiv},
-      primaryClass={physics.flu-dyn},
-      url={https://arxiv.org/abs/2410.14402}, 
+    author = {Crowe, Matthew. N. and Sutyrin, Georgi. G.},
+    title = {Symmetry breaking and nonlinear transformation of two-layer eastward propagating dipoles},
+    journal = {Physics of Fluids},
+    volume = {37},
+    number = {2},
+    pages = {021708},
+    year = {2025},
+    doi = {10.1063/5.0251761},
 }
