+++ paper/paper.bib
@@ -175,7 +175,7 @@
 }
 
 @misc{CroweSutyrin,
-    author = {Crowe, Matthew. N. and Sutyrin, Georgi. G.},
+    author = {Matthew N. Crowe and Georgi G. Sutyrin},
     title = {Symmetry breaking and nonlinear transformation of two-layer eastward propagating dipoles},
     journal = {Physics of Fluids},
     volume = {37},
