+++ paper/paper.bib
@@ -174,7 +174,7 @@
   url          = {https://github.com/mncrowe/QGDipoles.m}
 }
 
-@misc{CroweSutyrin,
+@article{CroweSutyrin,
     author = {Matthew N. Crowe and Georgi G. Sutyrin},
     title = {Symmetry breaking and nonlinear transformation of two-layer eastward propagating dipoles},
     journal = {Physics of Fluids},
