### 2025-04-04 #7767 @rwegener2
Hi @mncrowe, thanks for a really nice paper! It was a pleasure to read. A few comments for you:
* line 8 - Are there any assumptions in this package that restrict these solutions to the ocean, or would this code also work for atmospheric flows? Please specify if it is ocean only. If it can do both you might include stating that, but I'll leave it up to you.
* line 28 - use -> uses (or change modelling to models, whichever you prefer)
* line 31 - "transients" is a little bit jargon-y/shorthand to me. I suggest a slightly more descriptive phrase (non-physical transient features? transient features not indicative of realistic motion?)
* your first two paragraphs flow really nicely together and provide a nice background. thanks!
* line 35 - I struggled to follow the train of logic that leads to dipolar vortices. Are dipolar vortices an example of initial conditions that could solve the problems described in the previous paragraph?
* line 47 - Could you please explain what is meant by "It is designed to be consistent with the framework of Geophysicalflows.jl"? (ex. Does QGDipoles create outputs that can be used as inputs to Geophysicalflows? Do they create comparable outputes? etc. )
* line 54 - Personally I find "over long times" a little awkward. (maybe: over long timescales? stability of persistent 1- and 2-...? omit the phrase all together?) just a suggestion.
* line 56-58 - You've done a nice job keeping this paper accessible to an audience with a broad background. To keep that up, I suggest including a quick sentence or two stating the difference between a time-dependent GFD problem and a steady problem. This will help situate the uniqueness of your package
* the bulleted list of similar packages stating what they do and what they don't is very clear and easy to read!
* line 90 - Figure is stated twice because of the clickable Figure 1 link. Please remove the non-linked text "Figure"
* Figure 1 caption -- I assume these streamfunctions are shown left to right? I would specify in some way, either through figure titles, panel labels (ex. A -> C), or stating in the caption somehow (ex. "Lamb-Chaplygin dipole (left panel), ...")

Let me know if anything is unclear!
