### 2025-04-09 #7767 @mncrowe
> * line 8 - Are there any assumptions in this package that restrict these solutions to the ocean, or would this code also work for atmospheric flows? Please specify if it is ocean only. If it can do both you might include stating that, but I'll leave it up to you.

The equations and assumptions are valid in the atmosphere too and have been used, for example, to model atmospheric Rossby waves. They're also applicable to 2D flows in general (including non-rotating flow) as well as magnetohydrodynamical flows. As such, I'd like to keep this description general.

> * line 28 - use -> uses (or change modelling to models, whichever you prefer)

Changed to "uses".

> * line 31 - "transients" is a little bit jargon-y/shorthand to me. I suggest a slightly more descriptive phrase (non-physical transient features? transient features not indicative of realistic motion?)

Changed to "non-physical transient features". I've also attempted to clarify this idea by splitting the sentence and elaborating on how they're generated.

> * your first two paragraphs flow really nicely together and provide a nice background. thanks!

Thanks!

> * line 35 - I struggled to follow the train of logic that leads to dipolar vortices. Are dipolar vortices an example of initial conditions that could solve the problems described in the previous paragraph?

Yes, the idea is: QG solutions can solve this problem + dipolar vortices are common QG solutions => dipolar vortices are a potentially useful initial condition. I've attempted to clarify this with a new sentence at the end of the paragraph.

> * line 47 - Could you please explain what is meant by "It is designed to be consistent with the framework of Geophysicalflows.jl"? (ex. Does QGDipoles create outputs that can be used as inputs to Geophysicalflows? Do they create comparable outputes? etc. )

Geophysicalflows uses a `grid` structure which stores the physical and Fourier grids for the problem. QGDipoles can take this `grid` an use it to calculate a solution such that the solution can be used as an input for Geophysicalflows. So not only is the output consistent with the Geophysicalflows inputs, but they use equivalent `grid` definitions such that e.g. Fourier transforms obey the same conventions, fields have the same names, etc. I've tried to clarify this here but wanted to keep it brief.

> * line 54 - Personally I find "over long times" a little awkward. (maybe: over long timescales? stability of persistent 1- and 2-...? omit the phrase all together?) just a suggestion.

Changed to "very long timescales". This isn't very specific but anyone curious can see the paper. I just wanted to include this comment to point out that the package can be used for research.

> * line 56-58 - You've done a nice job keeping this paper accessible to an audience with a broad background. To keep that up, I suggest including a quick sentence or two stating the difference between a time-dependent GFD problem and a steady problem. This will help situate the uniqueness of your package

I've replaced "steady problem" with "steady-state solutions" which should be more familiar to readers. Let me know if you feel addition details is required.

> * the bulleted list of similar packages stating what they do and what they don't is very clear and easy to read!

Thanks!

> * line 90 - Figure is stated twice because of the clickable Figure 1 link. Please remove the non-linked text "Figure"

Fixed.

> * Figure 1 caption -- I assume these streamfunctions are shown left to right? I would specify in some way, either through figure titles, panel labels (ex. A -> C), or stating in the caption somehow (ex. "Lamb-Chaplygin dipole (left panel), ...")

Added left/centre/right.

> Let me know if anything is unclear!
