+++ paper/paper.md
@@ -34,25 +34,27 @@ The system is closed since the flow-field is related to the PV through 'PV inver
 QG systems are commonly used for idealised studies of large-scale ocean dynamics and are also some of the first equation sets studied by students in Geophysical Fluid dynamics courses across the world.
 
 QG dynamics break down when studying small-scale flows and in the presence of features that change the leading order force balance, such as coastal boundaries or steep topographic slopes.
-Therefore, realistic ocean-atmosphere modelling typically use 'primitive equation' models which solve variants of the Boussinesq Navier-Stokes equations.
-While more realistic, a major difficulty with these models can be setting initial conditions, with many models requiring long spin-up times to reach a state when transients generated by the initial evolution have decayed.
+Therefore, realistic ocean-atmosphere modelling typically uses 'primitive equation' models which solve variants of the Boussinesq Navier-Stokes equations.
+While more realistic, a major difficulty with these models can be setting initial conditions, with many models requiring long spin-up times to reach a state when non-physical transient features have decayed.
+These transient features arise as a consequence of the initial conditions not satisfying the full governing equations and boundary conditions and typically consist of a fast, wave-like response.
 Imposing initial conditions that obey QG dynamics can be a particularly effective strategy, as they exist in a 'balanced' state and do not generate fast inertial waves during the transient evolution phase.
 
 Dipolar vortices are propagating structures which consist of two counter-rotating monopolar vortices moving together via self-advection.
 Also known as Modons, they are commonly observed coherent structures [@NiZWH20] which are observed to be remarkably stable over long time scales [@NycanderI90].
 In the ocean, dipolar vortices can act to transport water masses over large distances while in atmospheric science they have been used to model nonlinear propagating (solitary) Rossby waves [@mcwilliams_1980; @rostami_zeitlin_2021].
 Much recent work has focussed on the long-time stability of dipolar vortices [@Davies_et_al] and they remain a topic of active research.
+The essential dynamics of dipolar vortices are well-captured by QG models hence these QG solutions may be used for generating 'balanced' initial conditions in realistic models in addition to their use in studies of QG systems.
 
 `QGDipoles.jl` provides a series of functions that allow users to easily generate dipolar vortex solutions to two of the most commonly used QG models; the layered QG system (LQG) and the surface QG (SQG) system.
 This package is intended both for those studying idealised vortex dynamics in a QG framework, and those looking to initialise a simulation with a steadily propagating, balanced vortex.
-It is designed to be consistent with the framework of `GeophysicalFlows.jl` [@GeophysicalFlows], a Julia package that contains modules for solving various QG systems, and accepts `grid` inputs generated using the `TwoDGrid` function from `FourierFlows.jl` [@FourierFlows].
-As such, `QGDipoles.jl` can generate solution arrays on both CPUs and GPUs using `CUDA.jl`.
+It is designed to be consistent with the framework of `GeophysicalFlows.jl` [@GeophysicalFlows], a Julia package that contains modules for solving various QG systems on CPUs and GPUs.
+As such, `QGDipoles.jl` can generate solution arrays on both CPUs and GPUs using `CUDA.jl`, accepts `grid` inputs generated using the `TwoDGrid` function from `FourierFlows.jl` [@FourierFlows] and produces outputs which can be directly passed to `GeophysicalFlows.jl` inputs.
 Full documentation exists with examples covering a range of LQG and SQG solutions.
-This package has been used in a recent study of the stability of 1- and 2-layer dipolar vortices over long times [@CroweSutyrin].
[…]
