+++ v1.0.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"Methodology/#Methodology","page":"Methodology","title":"Methodology","text":"","category":"section"},{"location":"Methodology/","page":"Methodology","title":"Methodology","text":"This page contains a summary of the numerical method used to determine steady dipolar solutions to quasi-geostrophic systems. A summary of the notation used is also included below.","category":"page"},{"location":"Methodology/#Method-Summary","page":"Methodology","title":"Method Summary","text":"","category":"section"},{"location":"Methodology/","page":"Methodology","title":"Methodology","text":"This package is based on the semi-analytic theory of dipolar vortices derived in Johnson & Crowe 2023[1] and Crowe & Johnson 2023[2] for SQG solutions and Crowe & Johnson 2024[3] for LQG solutions. A summary is presented here such that the notation and examples presented elsewhere make some sense. We consider a dipolar vortex of radius ell, moving with speed U. This vortex consists of an isolated region of high vorticity with a closed streamline at x^2 + y^2 = ell^2 (in a frame co-moving with the vortex), hence fluid does not escape during propagation. Within the vortex core, x^2 + y^2  ell^2, are two counter rotating regions, corresponding to a dipole. The streamfunction describing the flow is denoted by psi and potential vorticity (PV) anomaly by q. Velocities may be derived as (u v) = (-partial_ypsi partial_xpsi). The streamfunction, psi, and PV anomaly, q, are related through PV inversion. In the case of multiple layers, psi and q are vector valued functions of length equal to the number of layers, N.","category":"page"},{"location":"Methodology/#Layered-Quasi-Geostrophic-(LQG)-Solutions","page":"Methodology","title":"Layered Quasi-Geostrophic (LQG) Solutions","text":"","category":"section"},{"location":"Methodology/","page":"Methodology","title":"Methodology","text":"In the LQG model, steady, propagating, dipolar vortices satisfy the relation","category":"page"},{"location":"Methodology/","page":"Methodology","title":"Methodology","text":"q_i + beta_i y = F_i(psi_i + Uy)","category":"page"},{"location":"Methodology/","page":"Methodology","title":"Methodology","text":"where beta_i denotes the background PV gradient, i in 1dotsN is the layer index and F_i is an arbitrary (piecewise continuous) function. To proceed, we assume that F_i(z) = (beta_iU) z for x^2 + y^2  ell^2 (outside the vortex) and F_i(z) = -(K_i^2ell^2) z for x^2 + y^2  ell^2 (inside the vortex). Using a Hankel transform and expansion in term of Zernike radial functions, the problem may be reduced to the linear algebra system","category":"page"},{"location":"Methodology/","page":"Methodology","title":"Methodology","text":"left textbfA -  sum _n = 1^N K_n^2textbfB_n right textbfa = textbfc_0 + sum _n = 1^N K_n^2 textbfc_n","category":"page"},{"location":"Methodology/","page":"Methodology","title":"Methodology","text":"where textbfA and textbfB_n are matrices, textbfa is a vector containing the coefficients in the polynomial expansion, textbfc_j are vectors and the K_n are defined in F_i above and appear as unknown eigenvalues in the linear problem. In order to solve the system, N additional conditions are required. These are textbfd_n cdot textbfa = 0 for n in 1 dots N where the textbfd_n are vectors. These conditions correspond to the requirement that the streamfunction and vorticity are continuous in each layer. In principal, we have an infinite number of coefficients in textbfa. However, since we know that these coefficients must decay with increasing index (since psi is continuous), we can 
[…]
