### 2025-04-11 #7767 @rwegener2
Thanks for the responses, @mncrowe! The paper looks good.

For the last item:
* We generally request that the archive and the paper have the same titles and author lists. If there is an intentional reason they are different let me know, otherwise I'll ask that you please set the title and author lists to match.
