### 2025-04-14 #7767 @mncrowe
@rwegener2, I've updated the archive name and author lists. Can I keep the version name in the archive name? Or would it be better to remove it to match the paper?
