### 2025-04-14 #7767 @rwegener2
Thanks for updating the author lists.

For the title, it's up do you if you want to include the version number or not. The typical JOSS flow is that the archive and paper title match exactly, so in your case the title would be "QGDipoles.jl: A Julia package for calculating dipolar vortex solutions to the Quasi-Geostrophic equations". But if you'd like to include the version in there somewhere as well that's fine with me (ex. "QGDipoles.jl: A Julia package for calculating dipolar vortex solutions to the Quasi-Geostrophic equations (v1.0.1)"). I believe that Zenodo also lets you name versions, so you could name the archive without the version number but still associate the upload with a specific version.
