### 2025-04-17 #7767 @kthyng
@mncrowe Please check that your inline vs. parenthetical references are correct. You can compare with the example paper here: https://joss.readthedocs.io/en/latest/example_paper.html. I see some under "methodology".
